Migrate show view to ES6
