+++ package.json
@@ -18,5 +18,8 @@
     "sass": "^1.15.2",
     "sass-loader": "^8.0.0",
     "vue-template-compiler": "^2.6.11"
+  },
+  "dependencies": {
+    "ol": "^5.3.3"
   }
 }
