+++ src/public/assets/scripts/main.js
@@ -1,1 +1,1 @@
-!function(e){var t={};function n(i){if(t[i])return t[i].exports;var s=t[i]={i:i,l:!1,exports:{}};return e[i].call(s.exports,s,s.exports,n),s.l=!0,s.exports}n.m=e,n.c=t,n.d=function(e,t,i){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:i})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var i=Object.create(null);if(n.r(i),Object.defineProperty(i,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var s in e)n.d(i,s,function(t){return e[t]}.bind(null,s));return i},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="/",n(n.s=0)}({0:function(e,t,n){n("WfG0"),e.exports=n("zcrr")},WfG0:function(e,t,n){"use strict";n.r(t);var i=biigle.$require("core.mixins.loader"),s=biigle.$require("labelTrees.components.labelTypeahead"),o=biigle.$require("messages").handleErrorResponse,a={mixins:[i],components:{typeahead:s},data:{volumeId:null,useExistingAnnotations:!1,skipNoveltyDetection:!1,showAdvanced:!1,shouldFetchLabels:!1,labels:[],selectedLabels:[],submitted:!1},computed:{canSkipNoveltyDetection:function(){return this.useExistingAnnotations&&this.hasLabels},hasLabels:function(){return this.labels.length>0},hasSelectedLabels:function(){return this.selectedLabels.length>0},showRestrictLabelsInput:function(){return this.showAdvanced&&this.useExistingAnnotations&&this.hasLabels},hasNoExistingAnnotations:function(){return this.useExistingAnnotations&&!this.hasLabels&&!this.loading}},methods:{toggle:function(){this.showAdvanced=!this.showAdvanced},setLabels:function(e){this.labels=e.body},handleSelectedLabel:function(e){-1===this.selectedLabels.indexOf(e)&&this.selectedLabels.push(e)},handleUnselectLabel:function(e){var t=this.selectedLabels.indexOf(e);t>=0&&this.selectedLabels.splice(t,1)},submit:function(){this.submitted=!0}},watch:{useExistingAnnotations:function(e){e&&(this.shouldFetchLabels=!0)},shouldFetchLabels:function(e){e&&(this.startLoading(),this.$http.get("api/v1/volumes{/id}/annotation-labels",{params:{id:this.volumeId}}).then(this.setLabels,o).finally(this.finishLoading))}},created:function(){this.volumeId=biigle.$require("maia.volumeId"),this.useExistingAnnotations=biigle.$require("maia.useExistingAnnotations"),this.skipNoveltyDetection=biigle.$require("maia.skipNoveltyDetection"),this.useExistingAnnotations&&(this.shouldFetchLabels=!0)}};biigle.$mount("maia-job-form",a)},zcrr:function(e,t){}});
+!function(t){var e={};function n(i){if(e[i])return e[i].exports;var o=e[i]={i:i,l:!1,exports:{}};return t[i].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=t,n.c=e,n.d=function(t,e,i){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:i})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var i=Object.create(null);if(n.r(i),Object.defineProperty(i,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)n.d(i,o,function(e){return t[e]}.bind(null,o));return i},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="/",n(n.s=0)}({0:function(t,e,n){n("WfG0"),t.exports=n("zcrr")},"A1R+":function(t,e,n){"use strict";t.exports=o,t.exports.default=o;var i=n("YcpW");function o(t,e){if(!(this instanceof o))return new o(t,e);this._maxEntries=Math.max(4,t||9),this._minEntries=Math.max(2,Math.ceil(.4*this._maxEntries)),e&&this._initFormat(e),this.clear()}function s(t,e,n){if(!n)return e.indexOf(t);for(var i=0;i<e.length;i++)if(n(t,e[i]))return i;return-1}function r(t,e){a(t,0,t.children.length,e,t)}function a(t,e,n,i,o){o||(o=g(null)),o.minX=1/0,o.minY=1/0,o.maxX=-1/0,o.maxY=-1/0;for(var s,r=e;r<n;r++)s=t.children[r],h(o,t.leaf?i(s):s);return o}function h(t,e){return t.minX=Math.min(t.minX,e.minX),t.minY=Math.min(t.minY,e.minY),t.maxX=Math.max(t.maxX,e.maxX),t.maxY=Math.max(t.maxY,e.maxY),t}function u(t,e){return t.minX-e.minX}function c(t,e){return t.minY-e.minY}function d(t){return(t.maxX-t.minX)*(t.maxY-t.minY)}function l(t){return t.maxX-t.minX+(t.maxY-t.minY)}function p(t,e){return t.minX<=e.minX&&t.minY<=e.minY&&e.maxX<=t.maxX&&e.maxY<=t.maxY}function f(t,e){return e.minX<=t.maxX&&e.minY<=t.maxY&&e.maxX>=t.minX&&e.maxY>=t.minY}function g(t){return{children:t,height:1,leaf:!0,minX:1/0,minY:1/0,maxX:-1/0,maxY:-1/0}}function m(t,e,n,o,s){for(var r,a=[e,n];a.length;)(n=a.pop())-(e=a.pop())<=o||(r=e+Math.ceil((n-e)/o/2)*o,i(t,r,e,n,s),a.push(e,r,r,n))}o.prototype={all:function(){return this._all(this.data,[])},search:function(t){var e=this.data,n=[],i=this.toBBox;if(!f(t,e))return n;for(var o,s,r,a,h=[];e;){for(o=0,s=e.children.length;o<s;o++)r=e.children[o],f(t,a=e.leaf?i(r):r)&&(e.leaf?n.push(r):p(t,a)?this._all(r,n):h.push(r));e=h.pop()}return n},collides:function(t){var e=this.data,n=this.toBBox;if(!f(t,e))return!1;for(var i,o,s,r,a=[];e;){for(i=0,o=e.children.length;i<o;i++)if(s=e.children[i],f(t,r=e.leaf?n(s):s)){if(e.leaf||p(t,r))return!0;a.push(s)}e=a.pop()}return!1},load:function(t){if(!t||!t.length)return this;if(t.length<this._minEntries){for(var e=0,n=t.length;e<n;e++)this.insert(t[e]);return this}var i=this._build(t.slice(),0,t.length-1,0);if(this.data.children.length)if(this.data.height===i.height)this._splitRoot(this.data,i);else{if(this.data.height<i.height){var o=this.data;this.data=i,i=o}this._insert(i,this.data.height-i.height-1,!0)}else this.data=i;return this},insert:function(t){return t&&this._insert(t,this.data.height-1),this},clear:function(){return this.data=g([]),this},remove:function(t,e){if(!t)return this;for(var n,i,o,r,a=this.data,h=this.toBBox(t),u=[],c=[];a||u.length;){if(a||(a=u.pop(),i=u[u.length-1],n=c.pop(),r=!0),a.leaf&&-1!==(o=s(t,a.children,e)))return a.children.splice(o,1),u.push(a),this._condense(u),this;r||a.leaf||!p(a,h)?i?(n++,a=i.children[n],r=!1):a=null:(u.push(a),c.push(n),n=0,i=a,a=a.children[0])}return this},toBBox:function(t){return t},compareMinX:u,compareMinY:c,toJSON:function(){return this.data},fromJSON:function(t){return this.data=t,this},_all:function(t,e){for(var n=[];t;)t.leaf?e.push.apply(e,t.children):n.push.apply(n,t.children),t=n.pop();return e},_build:function(t,e,n,i){var o,s=n-e+1,a=this._maxEntries;if(s<=a)return r(o=g(t.slice(e,n+1)),this.toBBox),o;i||(i=Math.ceil(Math.log(s)/Math.log(a)),a=Math.ceil(s/Math.pow(a,i-1))),(o=g([])).leaf=!1,o.height=i;var h,u,c,d,l=Math.ceil(s/a),p=l*Math.ceil(Math.sqrt(a));for(m(t,e,n,p,this.compareMinX),h=e;h<=n;h+=p)for(m(t,h,c=Math.min(h+p-1,n),l,this.compareMinY),u=h;u<=c;u+=l)d=Math.min(u+l-1,c),o.children.push(this._build(t,u,d,i-1));return r(o,this.toBBox),o},_chooseSubtree:function(t,e,n,i){for(var o,s,r,a,h,u,c,l,p,f;i.push(e),!e.leaf&&i.length-1!==n;){for(c=l=1/0,o=0,s=e.children.length;o<s;o++)h=d(r=e.children[o]),p=t,f=r,(u=(Math.max(f.maxX,p.maxX)-Math.min(f.minX,p.minX))*(Math.max(f.maxY,p.maxY)-Math.min(f.minY,p.minY))-h)<l?(l=u,c=h<c?h:c,a=r):u===l&&h<c&&(c=h,a=r);e=a||e.children[0]}return e},_insert:function(t,e,n){var i=this.toBBox,o=n?t:i(t),s=[],r=this._chooseSubtree(o,this.data,e,s);for(r.children.push(t),h(r,o);e>=0&&s[e].children.length>this._maxEntries;)this._split(s,e),e--;this._adjustParentBBoxes(o,s,e)},_split:function(t,e){var n=t[e],i=n.children.length,o=this._minEntries;this._chooseSplitAxis(n,o,i);var s=this._chooseSplitIndex(n,o,i),a=g(n.children.splice(s,n.children.length-s));a.height=n.height,a.leaf=n.leaf,r(n,this.toBBox),r(a,this.toBBox),e?t[e-1].children.push(a):this._splitRoot(n,a)},_splitRoot:function(t,e){this.data=g([t,e]),this.data.height=t.height+1,this.data.leaf=!1,r(this.data,this.toBBox)},_chooseSplitIndex:function(t,e,n){var i,o,s,r,h,u,c,l,p,f,g,m,_,y;for(u=c=1/0,i=e;i<=n-e;i++)o=a(t,0,i,this.toBBox),s=a(t,i,n,this.toBBox),p=o,f=s,g=void 0,m=void 0,_=void 0,y=void 0,g=Math.max(p.minX,f.minX),m=Math.max(p.minY,f.minY),_=Math.min(p.maxX,f.maxX),y=Math.min(p.maxY,f.maxY),r=Math.max(0,_-g)*Math.max(0,y-m),h=d(o)+d(s),r<u?(u=r,l=i,c=h<c?h:c):r===u&&h<c&&(c=h,l=i);return l},_chooseSplitAxis:function(t,e,n){var i=t.leaf?this.compareMinX:u,o=t.leaf?this.compareMinY:c;this._allDistMargin(t,e,n,i)<this._allDistMargin(t,e,n,o)&&t.children.sort(i)},_allDistMargin:function(t,e,n,i){t.children.sort(i);var o,s,r=this.toBBox,u=a(t,0,e,r),c=a(t,n-e,n,r),d=l(u)+l(c);for(o=e;o<n-e;o++)s=t.children[o],h(u,t.leaf?r(s):s),d+=l(u);for(o=n-e-1;o>=e;o--)s=t.children[o],h(c,t.leaf?r(s):s),d+=l(c);return d},_adjustParentBBoxes:function(t,e,n){for(var i=n;i>=0;i--)h(e[i],t)},_condense:function(t){for(var e,n=t.length-1;n>=0;n--)0===t[n].children.length?n>0?(e=t[n-1].children).splice(e.indexOf(t[n]),1):this.clear():r(t[n],this.toBBox)},_initFormat:function(t){var e=["return a"," - b",";"];this.compareMinX=new Function("a","b",e.join(t[0])),this.compareMinY=new Function("a","b",e.join(t[1])),this.toBBox=new Function("a","return {minX: a"+t[0]+", minY: a"+t[1]+", maxX: a"+t[2]+", maxY: a"+t[3]+"};")}}},WfG0:function(t,e,n){"use strict";n.r(e);var i=biigle.$require("annotations.components.annotationCanvas"),o=biigle.$require("largo.mixins.annotationPatch"),s=biigle.$require("annotations.ol.AttachLabelInteraction"),r=biigle.$require("messages").handleErrorResponse,a=biigle.$require("volumes.components.imageGrid"),h=biigle.$require("volumes.components.imageGridImage"),u=biigle.$require("annotations.stores.images"),c=biigle.$require("keyboard"),d=biigle.$require("labelTrees.components.labelTrees"),l=biigle.$require("labelTrees.components.labelTypeahead"),p=biigle.$require("core.mixins.loader"),f=biigle.$require("messages"),g=biigle.$require("core.components.sidebar"),m=biigle.$require("core.components.sidebarTab"),_=biigle.$require("annotations.stores.styles"),y={mixins:[p],components:{typeahead:l},data:{volumeId:null,useExistingAnnotations:!1,skipNoveltyDetection:!1,showAdvanced:!1,shouldFetchLabels:!1,labels:[],selectedLabels:[],submitted:!1},computed:{canSkipNoveltyDetection:function(){return this.useExistingAnnotations&&this.hasLabels},hasLabels:function(){return this.labels.length>0},hasSelectedLabels:function(){return this.selectedLabels.length>0},showRestrictLabelsInput:function(){return this.showAdvanced&&this.useExistingAnnotations&&this.hasLabels},hasNoExistingAnnotations:function(){return this.useExistingAnnotations&&!this.hasLabels&&!this.loading}},methods:{toggle:function(){this.showAdvanced=!this.showAdvanced},setLabels:function(t){this.labels=t.body},handleSelectedLabel:function(t){-1===this.selectedLabels.indexOf(t)&&this.selectedLabels.push(t)},handleUnselectLabel:function(t){var e=this.selectedLabels.indexOf(t);e>=0&&this.selectedLabels.splice(e,1)},submit:function(){this.submitted=!0}},watch:{useExistingAnnotations:function(t){t&&(this.shouldFetchLabels=!0)},shouldFetchLabels:function(t){t&&(this.startLoading(),this.$http.get("api/v1/volumes{/id}/annotation-labels",{params:{id:this.volumeId}}).then(this.setLabels,r).finally(this.finishLoading))}},created:function(){this.volumeId=biigle.$require("maia.volumeId"),this.useExistingAnnotations=biigle.$require("maia.useExistingAnnotations"),this.skipNoveltyDetection=biigle.$require("maia.skipNoveltyDetection"),this.useExistingAnnotations&&(this.shouldFetchLabels=!0)}},v=Vue.resource("api/v1/maia/annotation-candidates{/id}"),I={mixins:[a],template:'<div class="image-grid" @wheel.prevent="scroll">\n        <div class="image-grid__images" ref="images">\n            <image-grid-image v-for="image in displayedImages" :key="image.id" :image="image" :empty-url="emptyUrl" :selectable="!isConverted(image)" :selected-icon="selectedIcon" @select="emitSelect"></image-grid-image>\n        </div>\n        <image-grid-progress v-if="canScroll" :progress="progress" @top="jumpToStart" @prev-page="reversePage" @prev-row="reverseRow" @jump="jumpToPercent" @next-row="advanceRow" @next-page="advancePage" @bottom="jumpToEnd"></image-grid-progress>\n    </div>',components:{imageGridImage:{mixins:[h,o],template:'<figure class="image-grid__image image-grid__image--annotation-candidate" :class="classObject" :title="title">\n        <div v-if="showIcon" class="image-icon">\n            <i class="fas" :class="iconClass"></i>\n        </div>\n        <img @click="toggleSelect" :src="url || emptyUrl" @error="showEmptyImage">\n        <div v-if="selected" class="attached-label">\n            <span class="attached-label__color" :style="labelStyle"></span>\n            <span class="attached-label__name" v-text="label.name"></span>\n        </div>\n    </figure>',computed:{label:function(){return this.selected?this.image.label:null},selected:function(){return this.$parent.isSelected(this.image)},converted:function(){return this.$parent.isConverted(this.image)},classObject:function(){return{"image-grid__image--selected":this.selected||this.converted,"image-grid__image--selectable":this.selectable,"image-grid__image--fade":this.selectedFade,"image-grid__image--small-icon":this.smallIcon}},iconClass:function(){return this.converted?"fa-lock":"fa-"+this.selectedIcon},showIcon:function(){return this.selectable||this.selected||this.converted},title:function(){return this.converted?"This annotation candidate has been converted":this.selected?"Detach label":"Attach selected label"},labelStyle:function(){return{"background-color":"#"+this.label.color}},id:function(){return this.image.id},uuid:function(){return this.image.uuid},urlTemplate:function(){return biigle.$require("maia.acUrlTemplate")}}}},props:{selectedCandidateIds:{type:Object,required:!0},convertedCandidateIds:{type:Object,required:!0}},methods:{isSelected:function(t){return this.selectedCandidateIds.hasOwnProperty(t.id)},isConverted:function(t){return this.convertedCandidateIds.hasOwnProperty(t.id)}}},x=Vue.resource("api/v1/maia-jobs{/id}",{},{save:{method:"POST",url:"api/v1/volumes{/id}/maia-jobs"},getTrainingProposals:{method:"GET",url:"api/v1/maia-jobs{/id}/training-proposals"},getTrainingProposalPoints:{method:"GET",url:"api/v1/maia-jobs{/jobId}/images{/imageId}/training-proposals"},getAnnotationCandidates:{method:"GET",url:"api/v1/maia-jobs{/id}/annotation-candidates"},getAnnotationCandidatePoints:{method:"GET",url:"api/v1/maia-jobs{/jobId}/images{/imageId}/annotation-candidates"},convertAnnotationCandidates:{method:"POST",url:"api/v1/maia-jobs{/id}/annotation-candidates"}}),C=Vue.resource("api/v1/maia/training-proposals{/id}"),b={mixins:[a],template:'<div class="image-grid" @wheel.prevent="scroll">\n        <div class="image-grid__images" ref="images">\n            <image-grid-image v-for="image in displayedImages" :key="image.id" :image="image" :empty-url="emptyUrl" :selectable="selectable" :selected-fade="selectable" :small-icon="!selectable" :selected-icon="selectedIcon" @select="emitSelect"></image-grid-image>\n        </div>\n        <image-grid-progress v-if="canScroll" :progress="progress" @top="jumpToStart" @prev-page="reversePage" @prev-row="reverseRow" @jump="jumpToPercent" @next-row="advanceRow" @next-page="advancePage" @bottom="jumpToEnd"></image-grid-progress>\n    </div>',components:{imageGridImage:{mixins:[h,o],template:'<figure class="image-grid__image" :class="classObject" :title="title">\n        <div v-if="showIcon" class="image-icon">\n            <i class="fas" :class="iconClass"></i>\n        </div>\n        <img @click="toggleSelect" :src="url || emptyUrl" @error="showEmptyImage">\n    </figure>',computed:{selected:function(){return this.$parent.selectedProposalIds.hasOwnProperty(this.image.id)},title:function(){if(this.selectable)return this.selected?"Unselect as interesting":"Select as interesting"},id:function(){return this.image.id},uuid:function(){return this.image.uuid},urlTemplate:function(){return biigle.$require("maia.tpUrlTemplate")}}}},props:{selectedProposalIds:{type:Object,required:!0}}};function P(){return function(){throw new Error("Unimplemented abstract method.")}()}var S=0;function w(t){return t.ol_uid||(t.ol_uid=String(++S))}var A=function(t){function e(e){var n="Assertion failed. See https://openlayers.org/en/"+("v"+"5.3.3".split("-")[0])+"/doc/errors/#"+e+" for details.";t.call(this,n),this.code=e,this.name="AssertionError",this.message=n}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(Error),E="add",F="remove",O="propertychange",T="function"==typeof Object.assign?Object.assign:function(t,e){var n=arguments;if(null==t)throw new TypeError("Cannot convert undefined or null to object");for(var i=Object(t),o=1,s=arguments.length;o<s;++o){var r=n[o];if(null!=r)for(var a in r)r.hasOwnProperty(a)&&(i[a]=r[a])}return i};function M(t){for(var e in t)delete t[e]}function L(t){var e;for(e in t)return!1;return!e}function k(t,e,n,i){for(var o,s=0,r=t.length;s<r;++s)if((o=t[s]).listener===e&&o.bindTo===n)return i&&(o.deleteIndex=s),o}function R(t,e){var n=j(t);return n?n[e]:void 0}function j(t,e){var n=t.ol_lm;return!n&&e&&(n=t.ol_lm={}),n}function D(t,e){var n=R(t,e);if(n){for(var i=0,o=n.length;i<o;++i)t.removeEventListener(e,n[i].boundListener),M(n[i]);n.length=0;var s=j(t);s&&(delete s[e],0===Object.keys(s).length&&function(t){delete t.ol_lm}(t))}}function G(t,e,n,i,o){var s=j(t,!0),r=s[e];r||(r=s[e]=[]);var a=k(r,n,i,!1);return a?o||(a.callOnce=!1):(a={bindTo:i,callOnce:!!o,listener:n,target:t,type:e},t.addEventListener(e,function(t){var e=function(e){var n=t.listener,i=t.bindTo||t.target;return t.callOnce&&W(t),n.call(i,e)};return t.boundListener=e,e}(a)),r.push(a)),a}function X(t,e,n,i){return G(t,e,n,i,!0)}function B(t,e,n,i){var o=R(t,e);if(o){var s=k(o,n,i,!0);s&&W(s)}}function W(t){if(t&&t.target){t.target.removeEventListener(t.type,t.boundListener);var e=R(t.target,t.type);if(e){var n="deleteIndex"in t?t.deleteIndex:e.indexOf(t);-1!==n&&e.splice(n,1),0===e.length&&D(t.target,t.type)}M(t)}}var q=function(){this.disposed_=!1};q.prototype.dispose=function(){this.disposed_||(this.disposed_=!0,this.disposeInternal())},q.prototype.disposeInternal=function(){};var Y=q;function $(){return!0}function z(){}var U=function(t){this.propagationStopped,this.type=t,this.target=null};U.prototype.preventDefault=function(){this.propagationStopped=!0},U.prototype.stopPropagation=function(){this.propagationStopped=!0};var V=U,N=function(t){function e(){t.call(this),this.pendingRemovals_={},this.dispatching_={},this.listeners_={}}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.addEventListener=function(t,e){var n=this.listeners_[t];n||(n=this.listeners_[t]=[]),-1===n.indexOf(e)&&n.push(e)},e.prototype.dispatchEvent=function(t){var e="string"==typeof t?new V(t):t,n=e.type;e.target=this;var i,o=this.listeners_[n];if(o){n in this.dispatching_||(this.dispatching_[n]=0,this.pendingRemovals_[n]=0),++this.dispatching_[n];for(var s=0,r=o.length;s<r;++s)if(!1===o[s].call(this,e)||e.propagationStopped){i=!1;break}if(--this.dispatching_[n],0===this.dispatching_[n]){var a=this.pendingRemovals_[n];for(delete this.pendingRemovals_[n];a--;)this.removeEventListener(n,z);delete this.dispatching_[n]}return i}},e.prototype.disposeInternal=function(){!function(t){var e=j(t);if(e)for(var n in e)D(t,n)}(this)},e.prototype.getListeners=function(t){return this.listeners_[t]},e.prototype.hasListener=function(t){return t?t in this.listeners_:Object.keys(this.listeners_).length>0},e.prototype.removeEventListener=function(t,e){var n=this.listeners_[t];if(n){var i=n.indexOf(e);t in this.pendingRemovals_?(n[i]=z,++this.pendingRemovals_[t]):(n.splice(i,1),0===n.length&&delete this.listeners_[t])}},e}(Y),K="change",J="clear";var H=function(t){function e(){t.call(this),this.revision_=0}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.changed=function(){++this.revision_,this.dispatchEvent(K)},e.prototype.getRevision=function(){return this.revision_},e.prototype.on=function(t,e){if(Array.isArray(t)){for(var n=t.length,i=new Array(n),o=0;o<n;++o)i[o]=G(this,t[o],e);return i}return G(this,t,e)},e.prototype.once=function(t,e){if(Array.isArray(t)){for(var n=t.length,i=new Array(n),o=0;o<n;++o)i[o]=X(this,t[o],e);return i}return X(this,t,e)},e.prototype.un=function(t,e){if(Array.isArray(t))for(var n=0,i=t.length;n<i;++n)B(this,t[n],e);else B(this,t,e)},e}(N),Z=function(t){function e(e,n,i){t.call(this,e),this.key=n,this.oldValue=i}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(V),Q={};function tt(t){return Q.hasOwnProperty(t)?Q[t]:Q[t]="change:"+t}var et=function(t){function e(e){t.call(this),w(this),this.values_={},void 0!==e&&this.setProperties(e)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.get=function(t){var e;return this.values_.hasOwnProperty(t)&&(e=this.values_[t]),e},e.prototype.getKeys=function(){return Object.keys(this.values_)},e.prototype.getProperties=function(){return T({},this.values_)},e.prototype.notify=function(t,e){var n;n=tt(t),this.dispatchEvent(new Z(n,t,e)),n=O,this.dispatchEvent(new Z(n,t,e))},e.prototype.set=function(t,e,n){if(n)this.values_[t]=e;else{var i=this.values_[t];this.values_[t]=e,i!==e&&this.notify(t,i)}},e.prototype.setProperties=function(t,e){for(var n in t)this.set(n,t[n],e)},e.prototype.unset=function(t,e){if(t in this.values_){var n=this.values_[t];delete this.values_[t],e||this.notify(t,n)}},e}(H),nt="length",it=function(t){function e(e,n){t.call(this,e),this.element=n}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(V),ot=function(t){function e(e,n){t.call(this);var i=n||{};if(this.unique_=!!i.unique,this.array_=e||[],this.unique_)for(var o=0,s=this.array_.length;o<s;++o)this.assertUnique_(this.array_[o],o);this.updateLength_()}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clear=function(){for(;this.getLength()>0;)this.pop()},e.prototype.extend=function(t){for(var e=0,n=t.length;e<n;++e)this.push(t[e]);return this},e.prototype.forEach=function(t){for(var e=this.array_,n=0,i=e.length;n<i;++n)t(e[n],n,e)},e.prototype.getArray=function(){return this.array_},e.prototype.item=function(t){return this.array_[t]},e.prototype.getLength=function(){return this.get(nt)},e.prototype.insertAt=function(t,e){this.unique_&&this.assertUnique_(e),this.array_.splice(t,0,e),this.updateLength_(),this.dispatchEvent(new it(E,e))},e.prototype.pop=function(){return this.removeAt(this.getLength()-1)},e.prototype.push=function(t){this.unique_&&this.assertUnique_(t);var e=this.getLength();return this.insertAt(e,t),this.getLength()},e.prototype.remove=function(t){for(var e=this.array_,n=0,i=e.length;n<i;++n)if(e[n]===t)return this.removeAt(n)},e.prototype.removeAt=function(t){var e=this.array_[t];return this.array_.splice(t,1),this.updateLength_(),this.dispatchEvent(new it(F,e)),e},e.prototype.setAt=function(t,e){var n=this.getLength();if(t<n){this.unique_&&this.assertUnique_(e,t);var i=this.array_[t];this.array_[t]=e,this.dispatchEvent(new it(F,i)),this.dispatchEvent(new it(E,e))}else{for(var o=n;o<t;++o)this.insertAt(o,void 0);this.insertAt(t,e)}},e.prototype.updateLength_=function(){this.set(nt,this.array_.length)},e.prototype.assertUnique_=function(t,e){for(var n=0,i=this.array_.length;n<i;++n)if(this.array_[n]===t&&n!==e)throw new A(58)},e}(et),st="VECTOR",rt="opacity",at="visible",ht="extent",ut="zIndex",ct="maxResolution",dt="minResolution",lt="source";function pt(t,e){if(!t)throw new A(e)}function ft(t,e,n){return Math.min(Math.max(t,e),n)}var gt="cosh"in Math?Math.cosh:function(t){var e=Math.exp(t);return(e+1/e)/2};var mt="precompose",_t="undefined",yt="ready";var vt=function(t){function e(e){var n=T({},e);delete n.source,t.call(this,n),this.mapPrecomposeKey_=null,this.mapRenderKey_=null,this.sourceChangeKey_=null,e.map&&this.setMap(e.map),G(this,tt(lt),this.handleSourcePropertyChange_,this);var i=e.source?e.source:null;this.setSource(i)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getLayersArray=function(t){var e=t||[];return e.push(this),e},e.prototype.getLayerStatesArray=function(t){var e=t||[];return e.push(this.getLayerState()),e},e.prototype.getSource=function(){return this.get(lt)||null},e.prototype.getSourceState=function(){var t=this.getSource();return t?t.getState():_t},e.prototype.handleSourceChange_=function(){this.changed()},e.prototype.handleSourcePropertyChange_=function(){this.sourceChangeKey_&&(W(this.sourceChangeKey_),this.sourceChangeKey_=null);var t=this.getSource();t&&(this.sourceChangeKey_=G(t,K,this.handleSourceChange_,this)),this.changed()},e.prototype.setMap=function(t){this.mapPrecomposeKey_&&(W(this.mapPrecomposeKey_),this.mapPrecomposeKey_=null),t||this.changed(),this.mapRenderKey_&&(W(this.mapRenderKey_),this.mapRenderKey_=null),t&&(this.mapPrecomposeKey_=G(t,mt,(function(t){var e=t,n=this.getLayerState();n.managed=!1,void 0===this.getZIndex()&&(n.zIndex=1/0),e.frameState.layerStatesArray.push(n),e.frameState.layerStates[w(this)]=n}),this),this.mapRenderKey_=G(this,K,t.render,t),this.changed())},e.prototype.setSource=function(t){this.set(lt,t)},e}(function(t){function e(e){t.call(this);var n=T({},e);n[rt]=void 0!==e.opacity?e.opacity:1,n[at]=void 0===e.visible||e.visible,n[ut]=e.zIndex,n[ct]=void 0!==e.maxResolution?e.maxResolution:1/0,n[dt]=void 0!==e.minResolution?e.minResolution:0,this.setProperties(n),this.state_=null,this.type}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getType=function(){return this.type},e.prototype.getLayerState=function(){var t=this.state_||{layer:this,managed:!0};return t.opacity=ft(this.getOpacity(),0,1),t.sourceState=this.getSourceState(),t.visible=this.getVisible(),t.extent=this.getExtent(),t.zIndex=this.getZIndex()||0,t.maxResolution=this.getMaxResolution(),t.minResolution=Math.max(this.getMinResolution(),0),this.state_=t,t},e.prototype.getLayersArray=function(t){return P()},e.prototype.getLayerStatesArray=function(t){return P()},e.prototype.getExtent=function(){return this.get(ht)},e.prototype.getMaxResolution=function(){return this.get(ct)},e.prototype.getMinResolution=function(){return this.get(dt)},e.prototype.getOpacity=function(){return this.get(rt)},e.prototype.getSourceState=function(){return P()},e.prototype.getVisible=function(){return this.get(at)},e.prototype.getZIndex=function(){return this.get(ut)},e.prototype.setExtent=function(t){this.set(ht,t)},e.prototype.setMaxResolution=function(t){this.set(ct,t)},e.prototype.setMinResolution=function(t){this.set(dt,t)},e.prototype.setOpacity=function(t){this.set(rt,t)},e.prototype.setVisible=function(t){this.set(at,t)},e.prototype.setZIndex=function(t){this.set(ut,t)},e}(et)),It="vector",xt=/^#([a-f0-9]{3}|[a-f0-9]{4}(?:[a-f0-9]{2}){0,2})$/i,Ct=/^([a-z]*)$/i;function bt(t){return"string"==typeof t?t:St(t)}!function(){var t={},e=0}();function Pt(t){return t[0]=ft(t[0]+.5|0,0,255),t[1]=ft(t[1]+.5|0,0,255),t[2]=ft(t[2]+.5|0,0,255),t[3]=ft(t[3],0,1),t}function St(t){var e=t[0];e!=(0|e)&&(e=e+.5|0);var n=t[1];n!=(0|n)&&(n=n+.5|0);var i=t[2];return i!=(0|i)&&(i=i+.5|0),"rgba("+e+","+n+","+i+","+(void 0===t[3]?1:t[3])+")"}function wt(t){return Array.isArray(t)?St(t):t}function At(t,e){var n=document.createElement("canvas");return t&&(n.width=t),e&&(n.height=e),n.getContext("2d")}var Et="undefined"!=typeof navigator?navigator.userAgent.toLowerCase():"",Ft=(Et.indexOf("firefox"),-1!==Et.indexOf("safari")&&Et.indexOf("chrom"),-1!==Et.indexOf("webkit")&&Et.indexOf("edge"),Et.indexOf("macintosh"),window.devicePixelRatio,function(){var t=!1;try{t=!!document.createElement("canvas").getContext("2d").setLineDash}catch(t){}return t}()),Ot=(navigator,window,window,navigator.msPointerEnabled,2),Tt=function(){var t,e={};return function(n){if(t||(t=document.createElement("div").style),!(n in e)){t.font=n;var i=t.fontFamily;if(t.font="",!i)return null;e[n]=i.split(/,\s?/)}return e[n]}}(),Mt=function(t){function e(e){t.call(this),this.highWaterMark=void 0!==e?e:2048,this.count_=0,this.entries_={},this.oldest_=null,this.newest_=null}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.canExpireCache=function(){return this.getCount()>this.highWaterMark},e.prototype.clear=function(){this.count_=0,this.entries_={},this.oldest_=null,this.newest_=null,this.dispatchEvent(J)},e.prototype.containsKey=function(t){return this.entries_.hasOwnProperty(t)},e.prototype.forEach=function(t,e){for(var n=this.oldest_;n;)t.call(e,n.value_,n.key_,this),n=n.newer},e.prototype.get=function(t){var e=this.entries_[t];return pt(void 0!==e,15),e===this.newest_||(e===this.oldest_?(this.oldest_=this.oldest_.newer,this.oldest_.older=null):(e.newer.older=e.older,e.older.newer=e.newer),e.newer=null,e.older=this.newest_,this.newest_.newer=e,this.newest_=e),e.value_},e.prototype.remove=function(t){var e=this.entries_[t];return pt(void 0!==e,15),e===this.newest_?(this.newest_=e.older,this.newest_&&(this.newest_.newer=null)):e===this.oldest_?(this.oldest_=e.newer,this.oldest_&&(this.oldest_.older=null)):(e.newer.older=e.older,e.older.newer=e.newer),delete this.entries_[t],--this.count_,e.value_},e.prototype.getCount=function(){return this.count_},e.prototype.getKeys=function(){var t,e=new Array(this.count_),n=0;for(t=this.newest_;t;t=t.older)e[n++]=t.key_;return e},e.prototype.getValues=function(){var t,e=new Array(this.count_),n=0;for(t=this.newest_;t;t=t.older)e[n++]=t.value_;return e},e.prototype.peekLast=function(){return this.oldest_.value_},e.prototype.peekLastKey=function(){return this.oldest_.key_},e.prototype.peekFirstKey=function(){return this.newest_.key_},e.prototype.pop=function(){var t=this.oldest_;return delete this.entries_[t.key_],t.newer&&(t.newer.older=null),this.oldest_=t.newer,this.oldest_||(this.newest_=null),--this.count_,t.value_},e.prototype.replace=function(t,e){this.get(t),this.entries_[t].value_=e},e.prototype.set=function(t,e){pt(!(t in this.entries_),16);var n={key_:t,newer:null,older:this.newest_,value_:e};this.newest_?this.newest_.newer=n:this.oldest_=n,this.newest_=n,this.entries_[t]=n,++this.count_},e.prototype.setSize=function(t){this.highWaterMark=t},e.prototype.prune=function(){for(;this.canExpireCache();)this.pop()},e}(N);new Array(6);var Lt=[0,0,0,1],kt=[0,0,0,1],Rt=new Mt,jt={},Dt=null,Gt={};!function(){var t,e,n=jt,i=["monospace","serif"],o=i.length,s="wmytzilWMYTZIL@#/&?$%10";function r(t){for(var n=Xt(),r=100;r<=700;r+=300){for(var a=r+" ",h=!0,u=0;u<o;++u){var c=i[u];if(n.font=a+"32px "+c,e=n.measureText(s).width,t!=c){n.font=a+"32px "+t+","+c;var d=n.measureText(s).width;h=h&&d!=e}}if(h)return!0}return!1}function a(){var e=!0;for(var i in n)n[i]<60&&(r(i)?(n[i]=60,M(Gt),Dt=null,Rt.clear()):(++n[i],e=!1));e&&(clearInterval(t),t=void 0)}}();function Xt(){return Dt||(Dt=At(1,1)),Dt}var Bt,Wt;Wt=Gt;var qt=function(t){this.opacity_=t.opacity,this.rotateWithView_=t.rotateWithView,this.rotation_=t.rotation,this.scale_=t.scale};qt.prototype.clone=function(){return new qt({opacity:this.getOpacity(),scale:this.getScale(),rotation:this.getRotation(),rotateWithView:this.getRotateWithView()})},qt.prototype.getOpacity=function(){return this.opacity_},qt.prototype.getRotateWithView=function(){return this.rotateWithView_},qt.prototype.getRotation=function(){return this.rotation_},qt.prototype.getScale=function(){return this.scale_},qt.prototype.getSnapToPixel=function(){return!1},qt.prototype.getAnchor=function(){return P()},qt.prototype.getImage=function(t){return P()},qt.prototype.getHitDetectionImage=function(t){return P()},qt.prototype.getImageState=function(){return P()},qt.prototype.getImageSize=function(){return P()},qt.prototype.getHitDetectionImageSize=function(){return P()},qt.prototype.getOrigin=function(){return P()},qt.prototype.getSize=function(){return P()},qt.prototype.setOpacity=function(t){this.opacity_=t},qt.prototype.setRotateWithView=function(t){this.rotateWithView_=t},qt.prototype.setRotation=function(t){this.rotation_=t},qt.prototype.setScale=function(t){this.scale_=t},qt.prototype.setSnapToPixel=function(t){},qt.prototype.listenImageChange=function(t,e){return P()},qt.prototype.load=function(){P()},qt.prototype.unlistenImageChange=function(t,e){P()};var Yt=function(t){function e(e){var n=e||{};t.call(this,{points:1/0,fill:n.fill,radius:n.radius,stroke:n.stroke,atlasManager:n.atlasManager})}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){var t=new e({fill:this.getFill()?this.getFill().clone():void 0,stroke:this.getStroke()?this.getStroke().clone():void 0,radius:this.getRadius(),atlasManager:this.atlasManager_});return t.setOpacity(this.getOpacity()),t.setScale(this.getScale()),t},e.prototype.setRadius=function(t){this.radius_=t,this.render_(this.atlasManager_)},e}(function(t){function e(e){var n=void 0!==e.rotateWithView&&e.rotateWithView;t.call(this,{opacity:1,rotateWithView:n,rotation:void 0!==e.rotation?e.rotation:0,scale:1}),this.checksums_=null,this.canvas_=null,this.hitDetectionCanvas_=null,this.fill_=void 0!==e.fill?e.fill:null,this.origin_=[0,0],this.points_=e.points,this.radius_=void 0!==e.radius?e.radius:e.radius1,this.radius2_=e.radius2,this.angle_=void 0!==e.angle?e.angle:0,this.stroke_=void 0!==e.stroke?e.stroke:null,this.anchor_=null,this.size_=null,this.imageSize_=null,this.hitDetectionImageSize_=null,this.atlasManager_=e.atlasManager,this.render_(this.atlasManager_)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){var t=new e({fill:this.getFill()?this.getFill().clone():void 0,points:this.getPoints(),radius:this.getRadius(),radius2:this.getRadius2(),angle:this.getAngle(),stroke:this.getStroke()?this.getStroke().clone():void 0,rotation:this.getRotation(),rotateWithView:this.getRotateWithView(),atlasManager:this.atlasManager_});return t.setOpacity(this.getOpacity()),t.setScale(this.getScale()),t},e.prototype.getAnchor=function(){return this.anchor_},e.prototype.getAngle=function(){return this.angle_},e.prototype.getFill=function(){return this.fill_},e.prototype.getHitDetectionImage=function(t){return this.hitDetectionCanvas_},e.prototype.getImage=function(t){return this.canvas_},e.prototype.getImageSize=function(){return this.imageSize_},e.prototype.getHitDetectionImageSize=function(){return this.hitDetectionImageSize_},e.prototype.getImageState=function(){return Ot},e.prototype.getOrigin=function(){return this.origin_},e.prototype.getPoints=function(){return this.points_},e.prototype.getRadius=function(){return this.radius_},e.prototype.getRadius2=function(){return this.radius2_},e.prototype.getSize=function(){return this.size_},e.prototype.getStroke=function(){return this.stroke_},e.prototype.listenImageChange=function(t,e){},e.prototype.load=function(){},e.prototype.unlistenImageChange=function(t,e){},e.prototype.render_=function(t){var e,n,i="",o="",s=0,r=null,a=0,h=0;this.stroke_&&(null===(n=this.stroke_.getColor())&&(n=kt),n=wt(n),void 0===(h=this.stroke_.getWidth())&&(h=1),r=this.stroke_.getLineDash(),a=this.stroke_.getLineDashOffset(),Ft||(r=null,a=0),void 0===(o=this.stroke_.getLineJoin())&&(o="round"),void 0===(i=this.stroke_.getLineCap())&&(i="round"),void 0===(s=this.stroke_.getMiterLimit())&&(s=10));var u=2*(this.radius_+h)+1,c={strokeStyle:n,strokeWidth:h,size:u,lineCap:i,lineDash:r,lineDashOffset:a,lineJoin:o,miterLimit:s};if(void 0===t){var d=At(u,u);this.canvas_=d.canvas,e=u=this.canvas_.width,this.draw_(c,d,0,0),this.createHitDetectionCanvas_(c)}else{u=Math.round(u);var l,p=!this.fill_;p&&(l=this.drawHitDetectionCanvas_.bind(this,c));var f=this.getChecksum(),g=t.add(f,u,u,this.draw_.bind(this,c),l);this.canvas_=g.image,this.origin_=[g.offsetX,g.offsetY],e=g.image.width,p?(this.hitDetectionCanvas_=g.hitImage,this.hitDetectionImageSize_=[g.hitImage.width,g.hitImage.height]):(this.hitDetectionCanvas_=this.canvas_,this.hitDetectionImageSize_=[e,e])}this.anchor_=[u/2,u/2],this.size_=[u,u],this.imageSize_=[e,e]},e.prototype.draw_=function(t,e,n,i){var o,s,r;e.setTransform(1,0,0,1,0,0),e.translate(n,i),e.beginPath();var a=this.points_;if(a===1/0)e.arc(t.size/2,t.size/2,this.radius_,0,2*Math.PI,!0);else{var h=void 0!==this.radius2_?this.radius2_:this.radius_;for(h!==this.radius_&&(a*=2),o=0;o<=a;o++)s=2*o*Math.PI/a-Math.PI/2+this.angle_,r=o%2==0?this.radius_:h,e.lineTo(t.size/2+r*Math.cos(s),t.size/2+r*Math.sin(s))}if(this.fill_){var u=this.fill_.getColor();null===u&&(u=Lt),e.fillStyle=wt(u),e.fill()}this.stroke_&&(e.strokeStyle=t.strokeStyle,e.lineWidth=t.strokeWidth,t.lineDash&&(e.setLineDash(t.lineDash),e.lineDashOffset=t.lineDashOffset),e.lineCap=t.lineCap,e.lineJoin=t.lineJoin,e.miterLimit=t.miterLimit,e.stroke()),e.closePath()},e.prototype.createHitDetectionCanvas_=function(t){if(this.hitDetectionImageSize_=[t.size,t.size],this.fill_)this.hitDetectionCanvas_=this.canvas_;else{var e=At(t.size,t.size);this.hitDetectionCanvas_=e.canvas,this.drawHitDetectionCanvas_(t,e,0,0)}},e.prototype.drawHitDetectionCanvas_=function(t,e,n,i){e.setTransform(1,0,0,1,0,0),e.translate(n,i),e.beginPath();var o=this.points_;if(o===1/0)e.arc(t.size/2,t.size/2,this.radius_,0,2*Math.PI,!0);else{var s,r,a,h=void 0!==this.radius2_?this.radius2_:this.radius_;for(h!==this.radius_&&(o*=2),s=0;s<=o;s++)a=2*s*Math.PI/o-Math.PI/2+this.angle_,r=s%2==0?this.radius_:h,e.lineTo(t.size/2+r*Math.cos(a),t.size/2+r*Math.sin(a))}e.fillStyle=bt(Lt),e.fill(),this.stroke_&&(e.strokeStyle=t.strokeStyle,e.lineWidth=t.strokeWidth,t.lineDash&&(e.setLineDash(t.lineDash),e.lineDashOffset=t.lineDashOffset),e.stroke()),e.closePath()},e.prototype.getChecksum=function(){var t=this.stroke_?this.stroke_.getChecksum():"-",e=this.fill_?this.fill_.getChecksum():"-";if(!this.checksums_||t!=this.checksums_[1]||e!=this.checksums_[2]||this.radius_!=this.checksums_[3]||this.radius2_!=this.checksums_[4]||this.angle_!=this.checksums_[5]||this.points_!=this.checksums_[6]){var n="r"+t+e+(void 0!==this.radius_?this.radius_.toString():"-")+(void 0!==this.radius2_?this.radius2_.toString():"-")+(void 0!==this.angle_?this.angle_.toString():"-")+(void 0!==this.points_?this.points_.toString():"-");this.checksums_=[n,t,e,this.radius_,this.radius2_,this.angle_,this.points_]}return this.checksums_[0]},e}(qt)),$t=function(t){var e=t||{};this.color_=void 0!==e.color?e.color:null,this.checksum_=void 0};$t.prototype.clone=function(){var t=this.getColor();return new $t({color:Array.isArray(t)?t.slice():t||void 0})},$t.prototype.getColor=function(){return this.color_},$t.prototype.setColor=function(t){this.color_=t,this.checksum_=void 0},$t.prototype.getChecksum=function(){if(void 0===this.checksum_){var t=this.color_;t?Array.isArray(t)||"string"==typeof t?this.checksum_="f"+bt(t):this.checksum_=w(this.color_):this.checksum_="f-"}return this.checksum_};var zt=$t,Ut=function(t){var e=t||{};this.color_=void 0!==e.color?e.color:null,this.lineCap_=e.lineCap,this.lineDash_=void 0!==e.lineDash?e.lineDash:null,this.lineDashOffset_=e.lineDashOffset,this.lineJoin_=e.lineJoin,this.miterLimit_=e.miterLimit,this.width_=e.width,this.checksum_=void 0};Ut.prototype.clone=function(){var t=this.getColor();return new Ut({color:Array.isArray(t)?t.slice():t||void 0,lineCap:this.getLineCap(),lineDash:this.getLineDash()?this.getLineDash().slice():void 0,lineDashOffset:this.getLineDashOffset(),lineJoin:this.getLineJoin(),miterLimit:this.getMiterLimit(),width:this.getWidth()})},Ut.prototype.getColor=function(){return this.color_},Ut.prototype.getLineCap=function(){return this.lineCap_},Ut.prototype.getLineDash=function(){return this.lineDash_},Ut.prototype.getLineDashOffset=function(){return this.lineDashOffset_},Ut.prototype.getLineJoin=function(){return this.lineJoin_},Ut.prototype.getMiterLimit=function(){return this.miterLimit_},Ut.prototype.getWidth=function(){return this.width_},Ut.prototype.setColor=function(t){this.color_=t,this.checksum_=void 0},Ut.prototype.setLineCap=function(t){this.lineCap_=t,this.checksum_=void 0},Ut.prototype.setLineDash=function(t){this.lineDash_=t,this.checksum_=void 0},Ut.prototype.setLineDashOffset=function(t){this.lineDashOffset_=t,this.checksum_=void 0},Ut.prototype.setLineJoin=function(t){this.lineJoin_=t,this.checksum_=void 0},Ut.prototype.setMiterLimit=function(t){this.miterLimit_=t,this.checksum_=void 0},Ut.prototype.setWidth=function(t){this.width_=t,this.checksum_=void 0},Ut.prototype.getChecksum=function(){return void 0===this.checksum_&&(this.checksum_="s",this.color_?"string"==typeof this.color_?this.checksum_+=this.color_:this.checksum_+=w(this.color_):this.checksum_+="-",this.checksum_+=","+(void 0!==this.lineCap_?this.lineCap_.toString():"-")+","+(this.lineDash_?this.lineDash_.toString():"-")+","+(void 0!==this.lineDashOffset_?this.lineDashOffset_:"-")+","+(void 0!==this.lineJoin_?this.lineJoin_:"-")+","+(void 0!==this.miterLimit_?this.miterLimit_.toString():"-")+","+(void 0!==this.width_?this.width_.toString():"-")),this.checksum_};var Vt=Ut,Nt=function(t){var e=t||{};this.geometry_=null,this.geometryFunction_=Ht,void 0!==e.geometry&&this.setGeometry(e.geometry),this.fill_=void 0!==e.fill?e.fill:null,this.image_=void 0!==e.image?e.image:null,this.renderer_=void 0!==e.renderer?e.renderer:null,this.stroke_=void 0!==e.stroke?e.stroke:null,this.text_=void 0!==e.text?e.text:null,this.zIndex_=e.zIndex};Nt.prototype.clone=function(){var t=this.getGeometry();return t&&"object"==typeof t&&(t=t.clone()),new Nt({geometry:t,fill:this.getFill()?this.getFill().clone():void 0,image:this.getImage()?this.getImage().clone():void 0,stroke:this.getStroke()?this.getStroke().clone():void 0,text:this.getText()?this.getText().clone():void 0,zIndex:this.getZIndex()})},Nt.prototype.getRenderer=function(){return this.renderer_},Nt.prototype.setRenderer=function(t){this.renderer_=t},Nt.prototype.getGeometry=function(){return this.geometry_},Nt.prototype.getGeometryFunction=function(){return this.geometryFunction_},Nt.prototype.getFill=function(){return this.fill_},Nt.prototype.setFill=function(t){this.fill_=t},Nt.prototype.getImage=function(){return this.image_},Nt.prototype.setImage=function(t){this.image_=t},Nt.prototype.getStroke=function(){return this.stroke_},Nt.prototype.setStroke=function(t){this.stroke_=t},Nt.prototype.getText=function(){return this.text_},Nt.prototype.setText=function(t){this.text_=t},Nt.prototype.getZIndex=function(){return this.zIndex_},Nt.prototype.setGeometry=function(t){"function"==typeof t?this.geometryFunction_=t:"string"==typeof t?this.geometryFunction_=function(e){return e.get(t)}:t?void 0!==t&&(this.geometryFunction_=function(){return t}):this.geometryFunction_=Ht,this.geometry_=t},Nt.prototype.setZIndex=function(t){this.zIndex_=t};var Kt=null;function Jt(t,e){if(!Kt){var n=new zt({color:"rgba(255,255,255,0.4)"}),i=new Vt({color:"#3399CC",width:1.25});Kt=[new Nt({image:new Yt({fill:n,stroke:i,radius:5}),fill:n,stroke:i})]}return Kt}function Ht(t){return t.getGeometry()}var Zt="renderOrder",Qt=function(t){function e(e){var n=e||{},i=T({},n);delete i.style,delete i.renderBuffer,delete i.updateWhileAnimating,delete i.updateWhileInteracting,t.call(this,i),this.declutter_=void 0!==n.declutter&&n.declutter,this.renderBuffer_=void 0!==n.renderBuffer?n.renderBuffer:100,this.style_=null,this.styleFunction_=void 0,this.setStyle(n.style),this.updateWhileAnimating_=void 0!==n.updateWhileAnimating&&n.updateWhileAnimating,this.updateWhileInteracting_=void 0!==n.updateWhileInteracting&&n.updateWhileInteracting,this.renderMode_=n.renderMode||It,this.type=st}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getDeclutter=function(){return this.declutter_},e.prototype.setDeclutter=function(t){this.declutter_=t},e.prototype.getRenderBuffer=function(){return this.renderBuffer_},e.prototype.getRenderOrder=function(){return this.get(Zt)},e.prototype.getStyle=function(){return this.style_},e.prototype.getStyleFunction=function(){return this.styleFunction_},e.prototype.getUpdateWhileAnimating=function(){return this.updateWhileAnimating_},e.prototype.getUpdateWhileInteracting=function(){return this.updateWhileInteracting_},e.prototype.setRenderOrder=function(t){this.set(Zt,t)},e.prototype.setStyle=function(t){this.style_=void 0!==t?t:Jt,this.styleFunction_=null===t?void 0:function(t){var e;if("function"==typeof t)e=t;else{var n;if(Array.isArray(t))n=t;else pt("function"==typeof t.getZIndex,41),n=[t];e=function(){return n}}return e}(this.style_),this.changed()},e.prototype.getRenderMode=function(){return this.renderMode_},e}(vt);Qt.prototype.getSource;var te=Qt;function ee(t,e,n,i,o){return o?(o[0]=t,o[1]=e,o[2]=n,o[3]=i,o):[t,e,n,i]}function ne(t,e){return t[0]==e[0]&&t[2]==e[2]&&t[1]==e[1]&&t[3]==e[3]}var ie="arraybuffer",oe="json",se="text",re="xml";function ae(t,e){return function(t,e,n,i){return function(o,s,r){var a=new XMLHttpRequest;a.open("GET","function"==typeof t?t(o,s,r):t,!0),e.getType()==ie&&(a.responseType="arraybuffer"),a.onload=function(t){if(!a.status||a.status>=200&&a.status<300){var o,s=e.getType();s==oe||s==se?o=a.responseText:s==re?(o=a.responseXML)||(o=(new DOMParser).parseFromString(a.responseText,"application/xml")):s==ie&&(o=a.response),o?n.call(this,e.readFeatures(o,{featureProjection:r}),e.readProjection(o),e.getLastExtent()):i.call(this)}else i.call(this)}.bind(this),a.onerror=function(){i.call(this)}.bind(this),a.send()}}(t,e,(function(t,e){"function"==typeof this.addFeatures&&this.addFeatures(t)}),z)}function he(t,e){return[[-1/0,-1/0,1/0,1/0]]}var ue={DEGREES:"degrees",FEET:"ft",METERS:"m",PIXELS:"pixels",TILE_PIXELS:"tile-pixels",USFEET:"us-ft"},ce={};ce[ue.DEGREES]=2*Math.PI*6370997/360,ce[ue.FEET]=.3048,ce[ue.METERS]=1,ce[ue.USFEET]=1200/3937;var de=ue,le=function(t){this.code_=t.code,this.units_=t.units,this.extent_=void 0!==t.extent?t.extent:null,this.worldExtent_=void 0!==t.worldExtent?t.worldExtent:null,this.axisOrientation_=void 0!==t.axisOrientation?t.axisOrientation:"enu",this.global_=void 0!==t.global&&t.global,this.canWrapX_=!(!this.global_||!this.extent_),this.getPointResolutionFunc_=t.getPointResolution,this.defaultTileGrid_=null,this.metersPerUnit_=t.metersPerUnit};le.prototype.canWrapX=function(){return this.canWrapX_},le.prototype.getCode=function(){return this.code_},le.prototype.getExtent=function(){return this.extent_},le.prototype.getUnits=function(){return this.units_},le.prototype.getMetersPerUnit=function(){return this.metersPerUnit_||ce[this.units_]},le.prototype.getWorldExtent=function(){return this.worldExtent_},le.prototype.getAxisOrientation=function(){return this.axisOrientation_},le.prototype.isGlobal=function(){return this.global_},le.prototype.setGlobal=function(t){this.global_=t,this.canWrapX_=!(!t||!this.extent_)},le.prototype.getDefaultTileGrid=function(){return this.defaultTileGrid_},le.prototype.setDefaultTileGrid=function(t){this.defaultTileGrid_=t},le.prototype.setExtent=function(t){this.extent_=t,this.canWrapX_=!(!this.global_||!t)},le.prototype.setWorldExtent=function(t){this.worldExtent_=t},le.prototype.setGetPointResolution=function(t){this.getPointResolutionFunc_=t},le.prototype.getPointResolutionFunc=function(){return this.getPointResolutionFunc_};var pe=le,fe=6378137*Math.PI,ge=[-fe,-fe,fe,fe],me=[-180,-85,180,85],_e=function(t){function e(e){t.call(this,{code:e,units:de.METERS,extent:ge,global:!0,worldExtent:me,getPointResolution:function(t,e){return t/gt(e[1]/6378137)}})}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(pe),ye=[new _e("EPSG:3857"),new _e("EPSG:102100"),new _e("EPSG:102113"),new _e("EPSG:900913"),new _e("urn:ogc:def:crs:EPSG:6.18:3:3857"),new _e("urn:ogc:def:crs:EPSG::3857"),new _e("http://www.opengis.net/gml/srs/epsg.xml#3857")];function ve(t,e,n){var i=t.length,o=n>1?n:2,s=e;void 0===s&&(s=o>2?t.slice():new Array(i));for(var r=fe,a=0;a<i;a+=o){s[a]=r*t[a]/180;var h=6378137*Math.log(Math.tan(Math.PI*(t[a+1]+90)/360));h>r?h=r:h<-r&&(h=-r),s[a+1]=h}return s}function Ie(t,e,n){var i=t.length,o=n>1?n:2,s=e;void 0===s&&(s=o>2?t.slice():new Array(i));for(var r=0;r<i;r+=o)s[r]=180*t[r]/fe,s[r+1]=360*Math.atan(Math.exp(t[r+1]/6378137))/Math.PI-90;return s}var xe=[-180,-90,180,90],Ce=6378137*Math.PI/180,be=function(t){function e(e,n){t.call(this,{code:e,units:de.DEGREES,extent:xe,axisOrientation:n,global:!0,metersPerUnit:Ce,worldExtent:xe})}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(pe),Pe=[new be("CRS:84"),new be("EPSG:4326","neu"),new be("urn:ogc:def:crs:EPSG::4326","neu"),new be("urn:ogc:def:crs:EPSG:6.6:4326","neu"),new be("urn:ogc:def:crs:OGC:1.3:CRS84"),new be("urn:ogc:def:crs:OGC:2:84"),new be("http://www.opengis.net/gml/srs/epsg.xml#4326","neu"),new be("urn:x-ogc:def:crs:EPSG:4326","neu")],Se={};var we,Ae,Ee,Fe={};function Oe(t,e,n){var i=t.getCode(),o=e.getCode();i in Fe||(Fe[i]={}),Fe[i][o]=n}function Te(t,e,n){var i;if(void 0!==e){for(var o=0,s=t.length;o<s;++o)e[o]=t[o];i=e}else i=t.slice();return i}function Me(t){!function(t,e){Se[t]=e}(t.getCode(),t),Oe(t,t,Te)}function Le(t){return"string"==typeof t?Se[t]||null:t||null}function ke(t){!function(t){t.forEach(Me)}(t),t.forEach((function(e){t.forEach((function(t){e!==t&&Oe(e,t,Te)}))}))}function Re(t){return t?Array.isArray(t)?function(e){return t}:"function"==typeof t?t:function(e){return[t]}:null}ke(ye),ke(Pe),we=ye,Ae=ve,Ee=Ie,Pe.forEach((function(t){we.forEach((function(e){Oe(t,e,Ae),Oe(e,t,Ee)}))}));var je=function(t){function e(e){t.call(this),this.projection_=Le(e.projection),this.attributions_=Re(e.attributions),this.attributionsCollapsible_=void 0===e.attributionsCollapsible||e.attributionsCollapsible,this.loading=!1,this.state_=void 0!==e.state?e.state:yt,this.wrapX_=void 0!==e.wrapX&&e.wrapX}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getAttributions=function(){return this.attributions_},e.prototype.getAttributionsCollapsible=function(){return this.attributionsCollapsible_},e.prototype.getProjection=function(){return this.projection_},e.prototype.getResolutions=function(){return P()},e.prototype.getState=function(){return this.state_},e.prototype.getWrapX=function(){return this.wrapX_},e.prototype.refresh=function(){this.changed()},e.prototype.setAttributions=function(t){this.attributions_=Re(t),this.changed()},e.prototype.setState=function(t){this.state_=t,this.changed()},e}(et),De="addfeature",Ge="changefeature",Xe="clear",Be="removefeature",We=n("A1R+"),qe=n.n(We),Ye=function(t){this.rbush_=qe()(t,void 0),this.items_={}};Ye.prototype.insert=function(t,e){var n={minX:t[0],minY:t[1],maxX:t[2],maxY:t[3],value:e};this.rbush_.insert(n),this.items_[w(e)]=n},Ye.prototype.load=function(t,e){for(var n=new Array(e.length),i=0,o=e.length;i<o;i++){var s=t[i],r=e[i],a={minX:s[0],minY:s[1],maxX:s[2],maxY:s[3],value:r};n[i]=a,this.items_[w(r)]=a}this.rbush_.load(n)},Ye.prototype.remove=function(t){var e=w(t),n=this.items_[e];return delete this.items_[e],null!==this.rbush_.remove(n)},Ye.prototype.update=function(t,e){var n=this.items_[w(e)];ne([n.minX,n.minY,n.maxX,n.maxY],t)||(this.remove(e),this.insert(t,e))},Ye.prototype.getAll=function(){return this.rbush_.all().map((function(t){return t.value}))},Ye.prototype.getInExtent=function(t){var e={minX:t[0],minY:t[1],maxX:t[2],maxY:t[3]};return this.rbush_.search(e).map((function(t){return t.value}))},Ye.prototype.forEach=function(t,e){return this.forEach_(this.getAll(),t,e)},Ye.prototype.forEachInExtent=function(t,e,n){return this.forEach_(this.getInExtent(t),e,n)},Ye.prototype.forEach_=function(t,e,n){for(var i,o=0,s=t.length;o<s;o++)if(i=e.call(n,t[o]))return i;return i},Ye.prototype.isEmpty=function(){return L(this.items_)},Ye.prototype.clear=function(){this.rbush_.clear(),this.items_={}},Ye.prototype.getExtent=function(t){var e=this.rbush_.toJSON();return ee(e.minX,e.minY,e.maxX,e.maxY,t)},Ye.prototype.concat=function(t){for(var e in this.rbush_.load(t.rbush_.all()),t.items_)this.items_[e]=t.items_[e]};var $e=Ye,ze=function(t){function e(e,n){t.call(this,e),this.feature=n}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(V),Ue=function(t){function e(e){var n=e||{};t.call(this,{attributions:n.attributions,projection:void 0,state:yt,wrapX:void 0===n.wrapX||n.wrapX}),this.loader_=z,this.format_=n.format,this.overlaps_=null==n.overlaps||n.overlaps,this.url_=n.url,void 0!==n.loader?this.loader_=n.loader:void 0!==this.url_&&(pt(this.format_,7),this.loader_=ae(this.url_,this.format_)),this.strategy_=void 0!==n.strategy?n.strategy:he;var i,o,s=void 0===n.useSpatialIndex||n.useSpatialIndex;this.featuresRtree_=s?new $e:null,this.loadedExtentsRtree_=new $e,this.nullGeometryFeatures_={},this.idIndex_={},this.undefIdIndex_={},this.featureChangeKeys_={},this.featuresCollection_=null,Array.isArray(n.features)?o=n.features:n.features&&(o=(i=n.features).getArray()),s||void 0!==i||(i=new ot(o)),void 0!==o&&this.addFeaturesInternal(o),void 0!==i&&this.bindFeaturesCollection_(i)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.addFeature=function(t){this.addFeatureInternal(t),this.changed()},e.prototype.addFeatureInternal=function(t){var e=w(t);if(this.addToIndex_(e,t)){this.setupChangeEvents_(e,t);var n=t.getGeometry();if(n){var i=n.getExtent();this.featuresRtree_&&this.featuresRtree_.insert(i,t)}else this.nullGeometryFeatures_[e]=t;this.dispatchEvent(new ze(De,t))}},e.prototype.setupChangeEvents_=function(t,e){this.featureChangeKeys_[t]=[G(e,K,this.handleFeatureChange_,this),G(e,O,this.handleFeatureChange_,this)]},e.prototype.addToIndex_=function(t,e){var n=!0,i=e.getId();return void 0!==i?i.toString()in this.idIndex_?n=!1:this.idIndex_[i.toString()]=e:(pt(!(t in this.undefIdIndex_),30),this.undefIdIndex_[t]=e),n},e.prototype.addFeatures=function(t){this.addFeaturesInternal(t),this.changed()},e.prototype.addFeaturesInternal=function(t){for(var e=[],n=[],i=[],o=0,s=t.length;o<s;o++){var r=t[o],a=w(r);this.addToIndex_(a,r)&&n.push(r)}for(var h=0,u=n.length;h<u;h++){var c=n[h],d=w(c);this.setupChangeEvents_(d,c);var l=c.getGeometry();if(l){var p=l.getExtent();e.push(p),i.push(c)}else this.nullGeometryFeatures_[d]=c}this.featuresRtree_&&this.featuresRtree_.load(e,i);for(var f=0,g=n.length;f<g;f++)this.dispatchEvent(new ze(De,n[f]))},e.prototype.bindFeaturesCollection_=function(t){var e=!1;G(this,De,(function(n){e||(e=!0,t.push(n.feature),e=!1)})),G(this,Be,(function(n){e||(e=!0,t.remove(n.feature),e=!1)})),G(t,E,(function(t){e||(e=!0,this.addFeature(t.element),e=!1)}),this),G(t,F,(function(t){e||(e=!0,this.removeFeature(t.element),e=!1)}),this),this.featuresCollection_=t},e.prototype.clear=function(t){if(t){for(var e in this.featureChangeKeys_){this.featureChangeKeys_[e].forEach(W)}this.featuresCollection_||(this.featureChangeKeys_={},this.idIndex_={},this.undefIdIndex_={})}else if(this.featuresRtree_)for(var n in this.featuresRtree_.forEach(this.removeFeatureInternal,this),this.nullGeometryFeatures_)this.removeFeatureInternal(this.nullGeometryFeatures_[n]);this.featuresCollection_&&this.featuresCollection_.clear(),this.featuresRtree_&&this.featuresRtree_.clear(),this.loadedExtentsRtree_.clear(),this.nullGeometryFeatures_={};var i=new ze(Xe);this.dispatchEvent(i),this.changed()},e.prototype.forEachFeature=function(t){if(this.featuresRtree_)return this.featuresRtree_.forEach(t);this.featuresCollection_&&this.featuresCollection_.forEach(t)},e.prototype.forEachFeatureAtCoordinateDirect=function(t,e){var n=[t[0],t[1],t[0],t[1]];return this.forEachFeatureInExtent(n,(function(n){return n.getGeometry().intersectsCoordinate(t)?e(n):void 0}))},e.prototype.forEachFeatureInExtent=function(t,e){if(this.featuresRtree_)return this.featuresRtree_.forEachInExtent(t,e);this.featuresCollection_&&this.featuresCollection_.forEach(e)},e.prototype.forEachFeatureIntersectingExtent=function(t,e){return this.forEachFeatureInExtent(t,(function(n){if(n.getGeometry().intersectsExtent(t)){var i=e(n);if(i)return i}}))},e.prototype.getFeaturesCollection=function(){return this.featuresCollection_},e.prototype.getFeatures=function(){var t;return this.featuresCollection_?t=this.featuresCollection_.getArray():this.featuresRtree_&&(t=this.featuresRtree_.getAll(),L(this.nullGeometryFeatures_)||function(t,e){for(var n=Array.isArray(e)?e:[e],i=n.length,o=0;o<i;o++)t[t.length]=n[o]}(t,function(t){var e=[];for(var n in t)e.push(t[n]);return e}(this.nullGeometryFeatures_))),t},e.prototype.getFeaturesAtCoordinate=function(t){var e=[];return this.forEachFeatureAtCoordinateDirect(t,(function(t){e.push(t)})),e},e.prototype.getFeaturesInExtent=function(t){return this.featuresRtree_.getInExtent(t)},e.prototype.getClosestFeatureToCoordinate=function(t,e){var n=t[0],i=t[1],o=null,s=[NaN,NaN],r=1/0,a=[-1/0,-1/0,1/0,1/0],h=e||$;return this.featuresRtree_.forEachInExtent(a,(function(t){if(h(t)){var e=t.getGeometry(),u=r;if((r=e.closestPointXY(n,i,s,r))<u){o=t;var c=Math.sqrt(r);a[0]=n-c,a[1]=i-c,a[2]=n+c,a[3]=i+c}}})),o},e.prototype.getExtent=function(t){return this.featuresRtree_.getExtent(t)},e.prototype.getFeatureById=function(t){var e=this.idIndex_[t.toString()];return void 0!==e?e:null},e.prototype.getFormat=function(){return this.format_},e.prototype.getOverlaps=function(){return this.overlaps_},e.prototype.getUrl=function(){return this.url_},e.prototype.handleFeatureChange_=function(t){var e=t.target,n=w(e),i=e.getGeometry();if(i){var o=i.getExtent();n in this.nullGeometryFeatures_?(delete this.nullGeometryFeatures_[n],this.featuresRtree_&&this.featuresRtree_.insert(o,e)):this.featuresRtree_&&this.featuresRtree_.update(o,e)}else n in this.nullGeometryFeatures_||(this.featuresRtree_&&this.featuresRtree_.remove(e),this.nullGeometryFeatures_[n]=e);var s=e.getId();if(void 0!==s){var r=s.toString();n in this.undefIdIndex_?(delete this.undefIdIndex_[n],this.idIndex_[r]=e):this.idIndex_[r]!==e&&(this.removeFromIdIndex_(e),this.idIndex_[r]=e)}else n in this.undefIdIndex_||(this.removeFromIdIndex_(e),this.undefIdIndex_[n]=e);this.changed(),this.dispatchEvent(new ze(Ge,e))},e.prototype.hasFeature=function(t){var e=t.getId();return void 0!==e?e in this.idIndex_:w(t)in this.undefIdIndex_},e.prototype.isEmpty=function(){return this.featuresRtree_.isEmpty()&&L(this.nullGeometryFeatures_)},e.prototype.loadFeatures=function(t,e,n){var i=this,o=this.loadedExtentsRtree_,s=this.strategy_(t,e);this.loading=!1;for(var r=function(t,r){var a=s[t];o.forEachInExtent(a,(function(t){return e=t.extent,n=a,e[0]<=n[0]&&n[2]<=e[2]&&e[1]<=n[1]&&n[3]<=e[3];var e,n}))||(i.loader_.call(i,a,e,n),o.insert(a,{extent:a.slice()}),i.loading=i.loader_!==z)},a=0,h=s.length;a<h;++a)r(a)},e.prototype.removeLoadedExtent=function(t){var e,n=this.loadedExtentsRtree_;n.forEachInExtent(t,(function(n){if(ne(n.extent,t))return e=n,!0})),e&&n.remove(e)},e.prototype.removeFeature=function(t){var e=w(t);e in this.nullGeometryFeatures_?delete this.nullGeometryFeatures_[e]:this.featuresRtree_&&this.featuresRtree_.remove(t),this.removeFeatureInternal(t),this.changed()},e.prototype.removeFeatureInternal=function(t){var e=w(t);this.featureChangeKeys_[e].forEach(W),delete this.featureChangeKeys_[e];var n=t.getId();void 0!==n?delete this.idIndex_[n.toString()]:delete this.undefIdIndex_[e],this.dispatchEvent(new ze(Be,t))},e.prototype.removeFromIdIndex_=function(t){var e=!1;for(var n in this.idIndex_)if(this.idIndex_[n]===t){delete this.idIndex_[n],e=!0;break}return e},e.prototype.setLoader=function(t){this.loader_=t},e}(je),Ve={mixins:[i],props:{unselectedAnnotations:{type:Array,default:function(){return[]}}},data:function(){return{selectingMaiaAnnotation:!1}},computed:{hasAnnotations:function(){return this.annotations.length>0}},methods:{handlePreviousImage:function(t){this.$emit("previous-image")},handleNextImage:function(t){this.$emit("next-image")},toggleSelectingMaiaAnnotation:function(){this.selectingMaiaAnnotation=!this.selectingMaiaAnnotation},createUnselectedAnnotationsLayer:function(){this.unselectedAnnotationFeatures=new ot,this.unselectedAnnotationSource=new Ue({features:this.unselectedAnnotationFeatures}),this.unselectedAnnotationLayer=new te({source:this.unselectedAnnotationSource,zIndex:99,updateWhileAnimating:!0,updateWhileInteracting:!0,style:_.editing,opacity:.5})},createSelectMaiaAnnotationInteraction:function(t){this.selectMaiaAnnotationInteraction=new s({map:this.map,features:t}),this.selectMaiaAnnotationInteraction.setActive(!1),this.selectMaiaAnnotationInteraction.on("attach",this.handleSelectMaiaAnnotation)},handleSelectMaiaAnnotation:function(t){this.$emit("select",t.feature.get("annotation"))},handleUnselectMaiaAnnotation:function(){this.selectedAnnotations.length>0&&this.$emit("unselect",this.selectedAnnotations[0])}},watch:{unselectedAnnotations:function(t){this.refreshAnnotationSource(t,this.unselectedAnnotationSource)},selectingMaiaAnnotation:function(t){this.selectMaiaAnnotationInteraction.setActive(t)}},created:function(){this.createUnselectedAnnotationsLayer(),this.map.addLayer(this.unselectedAnnotationLayer),this.selectInteraction.setActive(!1),this.canModify&&(this.createSelectMaiaAnnotationInteraction(this.unselectedAnnotationFeatures),this.map.addInteraction(this.selectMaiaAnnotationInteraction),c.on("Delete",this.handleUnselectMaiaAnnotation,0,this.listenerSet)),c.off("Shift+f",this.toggleMeasuring,this.listenerSet)},mounted:function(){c.off("m",this.toggleTranslating,this.listenerSet)}},Ne=[],Ke={},Je=[],He={},Ze={mixins:[p],components:{sidebar:g,sidebarTab:m,selectProposalsTab:{props:{proposals:{type:Array,required:!0},selectedProposals:{type:Array,required:!0}},computed:{selectedProposalsCount:function(){return this.selectedProposals.length},proposalsCount:function(){return this.proposals.length}},methods:{proceed:function(){this.$emit("proceed")}}},proposalsImageGrid:b,refineProposalsTab:{props:{selectedProposals:{type:Array,required:!0},seenProposals:{type:Array,required:!0}},computed:{numberSelectedProposals:function(){return this.selectedProposals.length},numberSeenProposals:function(){return this.seenProposals.length},hasNoSelectedProposals:function(){return 0===this.numberSelectedProposals},hasSeenAllSelectedProposals:function(){return this.numberSelectedProposals>0&&this.numberSelectedProposals===this.numberSeenProposals},textClass:function(){return this.hasSeenAllSelectedProposals?"text-success":""},buttonClass:function(){return this.hasSeenAllSelectedProposals?"btn-success":"btn-default"}}},refineCanvas:Ve,refineCandidatesCanvas:{mixins:[Ve],props:{convertedAnnotations:{type:Array,default:function(){return[]}}},methods:{createConvertedAnnotationsLayer:function(){this.convertedAnnotationFeatures=new ot,this.convertedAnnotationSource=new Ue({features:this.convertedAnnotationFeatures});var t=new et;t.set("color","999999"),this.convertedAnnotationLayer=new te({source:this.convertedAnnotationSource,zIndex:98,updateWhileAnimating:!0,updateWhileInteracting:!0,style:_.features(t)})}},watch:{convertedAnnotations:function(t){this.refreshAnnotationSource(t,this.convertedAnnotationSource)}},created:function(){this.createConvertedAnnotationsLayer(),this.map.addLayer(this.convertedAnnotationLayer)}},selectCandidatesTab:{components:{labelTrees:d},props:{labelTrees:{type:Array,required:!0}},methods:{handleSelectedLabel:function(t){this.$emit("select",t)},handleDeselectedLabel:function(t){this.$emit("select",null)},proceed:function(){this.$emit("proceed")}}},candidatesImageGrid:I,refineCandidatesTab:{components:{labelTrees:d},props:{selectedCandidates:{type:Array,required:!0},labelTrees:{type:Array,required:!0}},computed:{hasNoSelectedCandidates:function(){return 0===this.selectedCandidates.length}},methods:{handleSelectedLabel:function(t){this.$emit("select",t)},handleDeselectedLabel:function(t){this.$emit("select",null)},handleConvertCandidates:function(t){this.hasNoSelectedCandidates||this.$emit("convert")}}}},data:{job:null,states:null,labelTrees:[],visitedSelectProposalsTab:!1,visitedRefineProposalsTab:!1,visitedSelectCandidatesTab:!1,visitedRefineCandidatesTab:!1,openTab:"info",fetchProposalsPromise:null,hasProposals:!1,selectedProposalIds:{},seenProposalIds:{},lastSelectedProposal:null,currentProposalImage:null,currentProposalImageIndex:null,currentProposals:[],currentProposalsById:{},focussedProposal:null,proposalAnnotationCache:{},fetchCandidatesPromise:null,hasCandidates:!1,selectedCandidateIds:{},convertedCandidateIds:{},lastSelectedCandidate:null,currentCandidateImage:null,currentCandidateImageIndex:null,currentCandidates:[],currentCandidatesById:{},focussedCandidate:null,candidateAnnotationCache:{},selectedLabel:null,sequenceCounter:0},computed:{infoTabOpen:function(){return"info"===this.openTab},selectProposalsTabOpen:function(){return"select-proposals"===this.openTab},refineProposalsTabOpen:function(){return"refine-proposals"===this.openTab},selectCandidatesTabOpen:function(){return"select-candidates"===this.openTab},refineCandidatesTabOpen:function(){return"refine-candidates"===this.openTab},isInTrainingProposalState:function(){return this.job.state_id===this.states["training-proposals"]},isInAnnotationCandidateState:function(){return this.job.state_id===this.states["annotation-candidates"]},proposals:function(){return this.hasProposals?Ne:[]},selectedProposals:function(){var t=this.selectedProposalIds;return Object.keys(t).map((function(t){return Ke[t]})).sort((function(e,n){return e.image_id===n.image_id?t[e.id]-t[n.id]:e.image_id-n.image_id}))},proposalsForSelectView:function(){return this.isInTrainingProposalState?this.proposals:this.selectedProposals},hasSelectedProposals:function(){return this.selectedProposals.length>0},proposalImageIds:function(){var t={};return this.proposals.forEach((function(e){return t[e.image_id]=void 0})),Object.keys(t).map((function(t){return parseInt(t,10)}))},currentProposalImageId:function(){return this.proposalImageIds[this.currentProposalImageIndex]},nextProposalImageIndex:function(){return(this.currentProposalImageIndex+1)%this.proposalImageIds.length},nextProposalImageId:function(){return this.proposalImageIds[this.nextProposalImageIndex]},nextFocussedProposalImageId:function(){return this.nextFocussedProposal?this.nextFocussedProposal.image_id:this.nextProposalImageId},previousProposalImageIndex:function(){return(this.currentProposalImageIndex-1+this.proposalImageIds.length)%this.proposalImageIds.length},previousProposalImageId:function(){return this.proposalImageIds[this.previousProposalImageIndex]},hasCurrentProposalImage:function(){return null!==this.currentProposalImage},currentSelectedProposals:function(){var t=this;return this.currentProposals.filter((function(e){return t.selectedProposalIds.hasOwnProperty(e.id)}))},currentUnselectedProposals:function(){var t=this;return this.currentProposals.filter((function(e){return!t.selectedProposalIds.hasOwnProperty(e.id)}))},previousFocussedProposal:function(){var t=(this.selectedProposals.indexOf(this.focussedProposal)-1+this.selectedProposals.length)%this.selectedProposals.length;return this.selectedProposals[t]},nextFocussedProposal:function(){var t=(this.selectedProposals.indexOf(this.focussedProposal)+1)%this.selectedProposals.length;return this.selectedProposals[t]},focussedProposalToShow:function(){return this.refineProposalsTabOpen?this.focussedProposal:null},focussedProposalArray:function(){var t=this;return this.focussedProposalToShow?this.currentSelectedProposals.filter((function(e){return e.id===t.focussedProposalToShow.id})):[]},selectedAndSeenProposals:function(){var t=this;return this.selectedProposals.filter((function(e){return t.seenProposalIds.hasOwnProperty(e.id)}))},candidates:function(){return this.hasCandidates?Je:[]},selectedCandidates:function(){var t=this.selectedCandidateIds;return Object.keys(t).map((function(t){return He[t]})).sort((function(e,n){return e.image_id===n.image_id?t[e.id]-t[n.id]:e.image_id-n.image_id}))},hasSelectedCandidates:function(){return this.selectedCandidates.length>0},candidateImageIds:function(){var t={};return this.candidates.forEach((function(e){return t[e.image_id]=void 0})),Object.keys(t).map((function(t){return parseInt(t,10)}))},currentCandidateImageId:function(){return this.candidateImageIds[this.currentCandidateImageIndex]},nextCandidateImageIndex:function(){return(this.currentCandidateImageIndex+1)%this.candidateImageIds.length},nextCandidateImageId:function(){return this.candidateImageIds[this.nextCandidateImageIndex]},previousCandidateImageIndex:function(){return(this.currentCandidateImageIndex-1+this.candidateImageIds.length)%this.candidateImageIds.length},previousCandidateImageId:function(){return this.candidateImageIds[this.previousCandidateImageIndex]},hasCurrentCandidateImage:function(){return null!==this.currentCandidateImage},currentSelectedCandidates:function(){var t=this;return this.currentCandidates.filter((function(e){return t.selectedCandidateIds.hasOwnProperty(e.id)}))},currentUnselectedCandidates:function(){var t=this;return this.currentCandidates.filter((function(e){return!t.selectedCandidateIds.hasOwnProperty(e.id)&&!t.convertedCandidateIds.hasOwnProperty(e.id)}))},currentConvertedCandidates:function(){var t=this;return this.currentCandidates.filter((function(e){return t.convertedCandidateIds.hasOwnProperty(e.id)}))},previousFocussedCandidate:function(){var t=(this.selectedCandidates.indexOf(this.focussedCandidate)-1+this.selectedCandidates.length)%this.selectedCandidates.length;return this.selectedCandidates[t]},nextFocussedCandidate:function(){var t=(this.selectedCandidates.indexOf(this.focussedCandidate)+1)%this.selectedCandidates.length;return this.selectedCandidates[t]},nextFocussedCandidateImageId:function(){return this.nextFocussedCandidate?this.nextFocussedCandidate.image_id:this.nextCandidateImageId},focussedCandidateToShow:function(){return this.refineCandidatesTabOpen?this.focussedCandidate:null},focussedCandidateArray:function(){var t=this;return this.focussedCandidateToShow?this.currentSelectedCandidates.filter((function(e){return e.id===t.focussedCandidateToShow.id})):[]}},methods:{handleSidebarToggle:function(){var t=this;this.$nextTick((function(){t.$refs.proposalsImageGrid&&t.$refs.proposalsImageGrid.$emit("resize"),t.$refs.candidatesImageGrid&&t.$refs.candidatesImageGrid.$emit("resize")}))},handleTabOpened:function(t){this.openTab=t},setProposals:function(t){var e=this;(Ne=t.body).forEach((function(t){Ke[t.id]=t,e.setSelectedProposalId(t)})),this.hasProposals=Ne.length>0},fetchProposals:function(){return this.fetchProposalsPromise||(this.startLoading(),this.fetchProposalsPromise=x.getTrainingProposals({id:this.job.id}),this.fetchProposalsPromise.then(this.setProposals,r).finally(this.finishLoading)),this.fetchProposalsPromise},openRefineProposalsTab:function(){this.openTab="refine-proposals"},openRefineCandidatesTab:function(){this.openTab="refine-candidates"},updateSelectProposal:function(t,e){var n=this;t.selected=e,this.setSelectedProposalId(t);var i=C.update({id:t.id},{selected:e});return i.catch((function(i){r(i),t.selected=!e,n.setSelectedProposalId(t)})),i},setSelectedProposalId:function(t){t.selected?Vue.set(this.selectedProposalIds,t.id,this.getSequenceId()):Vue.delete(this.selectedProposalIds,t.id)},setSeenProposalId:function(t){Vue.set(this.seenProposalIds,t.id,!0)},fetchProposalAnnotations:function(t){return this.proposalAnnotationCache.hasOwnProperty(t)||(this.proposalAnnotationCache[t]=x.getTrainingProposalPoints({jobId:this.job.id,imageId:t}).then(this.parseAnnotations)),this.proposalAnnotationCache[t]},parseAnnotations:function(t){return Object.keys(t.body).map((function(e){return{id:parseInt(e,10),shape:"Circle",points:t.body[e]}}))},setCurrentProposalImageAndAnnotations:function(t){this.currentProposalImage=t[0],this.currentProposals=t[1],this.currentProposalsById={},this.currentProposals.forEach((function(t){this.currentProposalsById[t.id]=t}),this)},cacheNextProposalImage:function(){this.currentProposalImageId!==this.nextFocussedProposalImageId&&u.fetchImage(this.nextFocussedProposalImageId).catch((function(){}))},cacheNextProposalAnnotations:function(){this.currentProposalImageId!==this.nextFocussedProposalImageId&&this.fetchProposalAnnotations(this.nextFocussedProposalImageId).catch((function(){}))},handlePreviousProposalImage:function(){this.currentProposalImageIndex=this.previousProposalImageIndex},handlePreviousProposal:function(){this.previousFocussedProposal?this.focussedProposal=this.previousFocussedProposal:this.handlePreviousProposalImage()},handleNextProposalImage:function(){this.currentProposalImageIndex=this.nextProposalImageIndex},handleNextProposal:function(){this.nextFocussedProposal?this.focussedProposal=this.nextFocussedProposal:this.handleNextProposalImage()},handleRefineProposal:function(t){Vue.Promise.all(t.map(this.updateProposalPoints)).catch(r)},updateProposalPoints:function(t){var e=this.currentProposalsById[t.id];return C.update({id:t.id},{points:t.points}).then((function(){e.points=t.points}))},focusProposalToShow:function(){if(this.focussedProposalToShow){var t=this.currentProposalsById[this.focussedProposalToShow.id];t&&this.$refs.refineProposalsCanvas.focusAnnotation(t,!0,!1)}},handleSelectedProposal:function(t,e){t.selected?this.unselectProposal(t):e.shiftKey&&this.lastSelectedProposal?this.doForEachBetween(this.proposals,t,this.lastSelectedProposal,this.selectProposal):(this.lastSelectedProposal=t,this.selectProposal(t))},selectProposal:function(t){this.updateSelectProposal(t,!0).then(this.maybeInitFocussedProposal)},unselectProposal:function(t){var e=this.nextFocussedProposal;this.updateSelectProposal(t,!1).bind(this).then((function(){this.maybeUnsetFocussedProposal(t,e)}))},maybeInitFocussedProposal:function(){!this.focussedProposal&&this.hasSelectedProposals&&(this.focussedProposal=this.selectedProposals[0])},maybeUnsetFocussedProposal:function(t,e){this.focussedProposal&&this.focussedProposal.id===t.id&&(e&&e.id!==t.id?this.focussedProposal=e:this.focussedProposal=null)},maybeInitCurrentProposalImage:function(){null===this.currentProposalImageIndex&&(this.currentProposalImageIndex=0)},maybeInitCurrentCandidateImage:function(){null===this.currentCandidateImageIndex&&(this.currentCandidateImageIndex=0)},handleLoadingError:function(t){f.danger(t)},setSelectedCandidateId:function(t){t.label&&!t.annotation_id?Vue.set(this.selectedCandidateIds,t.id,this.getSequenceId()):Vue.delete(this.selectedCandidateIds,t.id)},setConvertedCandidateId:function(t){t.annotation_id?Vue.set(this.convertedCandidateIds,t.id,t.annotation_id):Vue.delete(this.convertedCandidateIds,t.id)},setCandidates:function(t){(Je=t.body).forEach((function(t){He[t.id]=t,this.setSelectedCandidateId(t),this.setConvertedCandidateId(t)}),this),this.hasCandidates=Je.length>0},fetchCandidates:function(){return this.fetchCandidatesPromise||(this.startLoading(),this.fetchCandidatesPromise=x.getAnnotationCandidates({id:this.job.id}),this.fetchCandidatesPromise.then(this.setCandidates,r).finally(this.finishLoading)),this.fetchCandidatesPromise},handleSelectedCandidate:function(t,e){t.label?this.unselectCandidate(t):e.shiftKey&&this.lastSelectedCandidate&&this.selectedLabel?this.doForEachBetween(this.candidates,t,this.lastSelectedCandidate,this.selectCandidate):(this.lastSelectedCandidate=t,this.selectCandidate(t))},selectCandidate:function(t){this.selectedLabel?t.annotation_id||this.updateSelectCandidate(t,this.selectedLabel).then(this.maybeInitFocussedCandidate):f.info("Please select a label first.")},unselectCandidate:function(t){var e=this.nextFocussedCandidate;this.updateSelectCandidate(t,null).bind(this).then((function(){this.maybeUnsetFocussedCandidate(t,e)}))},updateSelectCandidate:function(t,e){var n=this,i=t.label;t.label=e,this.setSelectedCandidateId(t);var o=e?e.id:null,s=v.update({id:t.id},{label_id:o});return s.catch((function(e){r(e),t.label=i,n.setSelectedCandidateId(t)})),s},fetchCandidateAnnotations:function(t){return this.candidateAnnotationCache.hasOwnProperty(t)||(this.candidateAnnotationCache[t]=x.getAnnotationCandidatePoints({jobId:this.job.id,imageId:t}).then(this.parseAnnotations)),this.candidateAnnotationCache[t]},setCurrentCandidateImageAndAnnotations:function(t){var e=this;this.currentCandidateImage=t[0],this.currentCandidates=t[1],this.currentCandidatesById={},this.currentCandidates.forEach((function(t){return e.currentCandidatesById[t.id]=t}))},handlePreviousCandidateImage:function(){this.currentCandidateImageIndex=this.previousCandidateImageIndex},handlePreviousCandidate:function(){this.previousFocussedCandidate?this.focussedCandidate=this.previousFocussedCandidate:this.handlePreviousCandidateImage()},handleNextCandidateImage:function(){this.currentCandidateImageIndex=this.nextCandidateImageIndex},handleNextCandidate:function(){this.nextFocussedCandidate?this.focussedCandidate=this.nextFocussedCandidate:this.handleNextCandidateImage()},focusCandidateToShow:function(){if(this.focussedCandidateToShow){var t=this.currentCandidatesById[this.focussedCandidateToShow.id];t&&this.$refs.refineCandidatesCanvas.focusAnnotation(t,!0,!1)}},maybeInitFocussedCandidate:function(){!this.focussedCandidate&&this.hasSelectedCandidates&&(this.focussedCandidate=this.selectedCandidates[0])},maybeUnsetFocussedCandidate:function(t,e){this.focussedCandidate&&this.focussedCandidate.id===t.id&&(e&&e.id!==t.id?this.focussedCandidate=e:this.focussedCandidate=null)},handleSelectedLabel:function(t){this.selectedLabel=t},doForEachBetween:function(t,e,n,i){var o=t.indexOf(e),s=t.indexOf(n);if(s<o){var r=s;s=o,o=r+1}else s-=1;for(var a=o;a<=s;a++)i(t[a])},cacheNextCandidateImage:function(){this.currentCandidateImageId!==this.nextFocussedCandidateImageId&&u.fetchImage(this.nextFocussedCandidateImageId).catch((function(){}))},cacheNextCandidateAnnotations:function(){this.currentCandidateImageId!==this.nextFocussedCandidateImageId&&this.fetchCandidateAnnotations(this.nextFocussedCandidateImageId).catch((function(){}))},handleRefineCandidate:function(t){Vue.Promise.all(t.map(this.updateCandidatePoints)).catch(r)},updateCandidatePoints:function(t){var e=this.currentCandidatesById[t.id];return v.update({id:t.id},{points:t.points}).then((function(){return e.points=t.points}))},handleConvertCandidates:function(){this.startLoading(),x.convertAnnotationCandidates({id:this.job.id},{}).then(this.handleConvertedCandidates,r).finally(this.finishLoading)},handleConvertedCandidates:function(t){var e=this;Object.keys(t.body).forEach((function(n){var i=e.nextFocussedCandidate,o=He[n];o.annotation_id=t.body[n],e.setSelectedCandidateId(o),e.setConvertedCandidateId(o),e.maybeUnsetFocussedCandidate(o,i)}))},getSequenceId:function(){return this.sequenceCounter++}},watch:{selectProposalsTabOpen:function(t){this.visitedSelectProposalsTab=!0,t&&c.setActiveSet("select-proposals")},refineProposalsTabOpen:function(t){this.visitedRefineProposalsTab=!0,t&&(c.setActiveSet("refine-proposals"),this.maybeInitFocussedProposal())},selectCandidatesTabOpen:function(t){this.visitedSelectCandidatesTab=!0,t&&c.setActiveSet("select-candidates")},refineCandidatesTabOpen:function(t){this.visitedRefineCandidatesTab=!0,t&&c.setActiveSet("refine-candidates")},visitedSelectProposalsTab:function(){this.fetchProposals()},visitedRefineProposalsTab:function(){this.fetchProposals().then(this.maybeInitFocussedProposal).then(this.maybeInitCurrentProposalImage)},visitedSelectCandidatesTab:function(){this.fetchCandidates()},visitedRefineCandidatesTab:function(){this.fetchCandidates().then(this.maybeInitFocussedCandidate).then(this.maybeInitCurrentCandidateImage)},currentProposalImageId:function(t){t?(this.startLoading(),Vue.Promise.all([u.fetchAndDrawImage(t),this.fetchProposalAnnotations(t),this.fetchProposals()]).then(this.setCurrentProposalImageAndAnnotations).then(this.focusProposalToShow).then(this.cacheNextProposalImage).then(this.cacheNextProposalAnnotations).catch(this.handleLoadingError).finally(this.finishLoading)):this.setCurrentProposalImageAndAnnotations([null,null])},focussedProposalToShow:function(t,e){t&&(e&&e.image_id===t.image_id?this.focusProposalToShow():this.currentProposalImageIndex=this.proposalImageIds.indexOf(t.image_id),this.setSeenProposalId(t))},currentCandidateImageId:function(t){t?(this.startLoading(),Vue.Promise.all([u.fetchAndDrawImage(t),this.fetchCandidateAnnotations(t),this.fetchCandidates()]).then(this.setCurrentCandidateImageAndAnnotations).then(this.focusCandidateToShow).then(this.cacheNextCandidateImage).then(this.cacheNextCandidateAnnotations).catch(this.handleLoadingError).finally(this.finishLoading)):this.setCurrentCandidateImageAndAnnotations([null,null])},focussedCandidateToShow:function(t,e){t&&(e&&e.image_id===t.image_id?this.focusCandidateToShow():this.currentCandidateImageIndex=this.candidateImageIds.indexOf(t.image_id))}},created:function(){this.job=biigle.$require("maia.job"),this.states=biigle.$require("maia.states"),this.labelTrees=biigle.$require("maia.labelTrees")}};biigle.$mount("maia-job-form",y),biigle.$mount("maia-show-container",Ze)},YcpW:function(t,e,n){t.exports=function(){"use strict";function t(t,e,n){var i=t[e];t[e]=t[n],t[n]=i}function e(t,e){return t<e?-1:t>e?1:0}return function(n,i,o,s,r){!function e(n,i,o,s,r){for(;s>o;){if(s-o>600){var a=s-o+1,h=i-o+1,u=Math.log(a),c=.5*Math.exp(2*u/3),d=.5*Math.sqrt(u*c*(a-c)/a)*(h-a/2<0?-1:1),l=Math.max(o,Math.floor(i-h*c/a+d)),p=Math.min(s,Math.floor(i+(a-h)*c/a+d));e(n,i,l,p,r)}var f=n[i],g=o,m=s;for(t(n,o,i),r(n[s],f)>0&&t(n,o,s);g<m;){for(t(n,g,m),g++,m--;r(n[g],f)<0;)g++;for(;r(n[m],f)>0;)m--}0===r(n[o],f)?t(n,o,m):(m++,t(n,m,s)),m<=i&&(o=m+1),i<=m&&(s=m-1)}}(n,i,o||0,s||n.length-1,r||e)}}()},zcrr:function(t,e){}});
