+++ src/public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/assets/scripts/main.js": "/assets/scripts/main.js?id=24be0c098b67e111b172",
+    "/assets/scripts/main.js": "/assets/scripts/main.js?id=a10c682ebf60003bd422",
     "/assets/styles/main.css": "/assets/styles/main.css?id=e0eebb85d62abf60780d"
 }
