+++ src/resources/assets/js/components/candidatesImageGrid.js
@@ -1,12 +1,21 @@
+import Image from './candidatesImageGridImage';
+import {ImageGrid} from '../import';
+
 /**
  * A letiant of the image grid used for the selection of MAIA annotation candidates.
  *
  * @type {Object}
  */
-biigle.$component('maia.components.candidatesImageGrid', {
-    mixins: [biigle.$require('volumes.components.imageGrid')],
+export default {
+    mixins: [ImageGrid],
+    template: `<div class="image-grid" @wheel.prevent="scroll">
+        <div class="image-grid__images" ref="images">
+            <image-grid-image v-for="image in displayedImages" :key="image.id" :image="image" :empty-url="emptyUrl" :selectable="!isConverted(image)" :selected-icon="selectedIcon" @select="emitSelect"></image-grid-image>
+        </div>
+        <image-grid-progress v-if="canScroll" :progress="progress" @top="jumpToStart" @prev-page="reversePage" @prev-row="reverseRow" @jump="jumpToPercent" @next-row="advanceRow" @next-page="advancePage" @bottom="jumpToEnd"></image-grid-progress>
+    </div>`,
     components: {
-        imageGridImage: biigle.$require('maia.components.candidatesImageGridImage'),
+        imageGridImage: Image,
     },
     props: {
         selectedCandidateIds: {
@@ -18,4 +27,12 @@ biigle.$component('maia.components.candidatesImageGrid', {
             required: true,
         },
     },
-});
+    methods: {
+        isSelected(image) {
+            return this.selectedCandidateIds.hasOwnProperty(image.id);
+        },
+        isConverted(image) {
+            return this.convertedCandidateIds.hasOwnProperty(image.id);
+        },
+    },
+};
