+++ src/resources/assets/js/components/candidatesImageGridImage.js
@@ -1,47 +1,39 @@
+import {AnnotationPatch} from '../import';
+import {ImageGridImage} from '../import';
+
 /**
  * A letiant of the image grid image used for the selection of MAIA annotation candidates.
  *
  * @type {Object}
  */
-biigle.$component('maia.components.candidatesImageGridImage', {
+export default {
     mixins: [
-        biigle.$require('volumes.components.imageGridImage'),
-        biigle.$require('largo.mixins.annotationPatch'),
+        ImageGridImage,
+        AnnotationPatch,
     ],
-    template: '<figure class="image-grid__image image-grid__image--annotation-candidate" :class="classObject" :title="title">' +
-        '<div v-if="showIcon" class="image-icon">' +
-            '<i class="fas" :class="iconClass"></i>' +
-        '</div>' +
-        '<img @click="toggleSelect" :src="url || emptyUrl" @error="showEmptyImage">' +
-        '<div v-if="showAnnotationLink" class="image-buttons">' +
-            '<a :href="showAnnotationLink" target="_blank" class="image-button" title="Show the annotation in the annotation tool">' +
-                '<span class="fa fa-external-link-square-alt" aria-hidden="true"></span>' +
-            '</a>' +
-        '</div>' +
-        '<div v-if="selected" class="attached-label">' +
-            '<span class="attached-label__color" :style="labelStyle"></span> ' +
-            '<span class="attached-label__name" v-text="label.name"></span>' +
-        '</div>' +
-    '</figure>',
+    template: `<figure class="image-grid__image image-grid__image--annotation-candidate" :class="classObject" :title="title">
+        <div v-if="showIcon" class="image-icon">
+            <i class="fas" :class="iconClass"></i>
+        </div>
+        <img @click="toggleSelect" :src="url || emptyUrl" @error="showEmptyImage">
+        <div v-if="selected" class="attached-label">
+            <span class="attached-label__color" :style="labelStyle"></span>
+            <span class="attached-label__name" v-text="label.name"></span>
+        </div>
+    </figure>`,
     computed: {
-        showAnnotationLink() {
-            return false;
-        },
         label() {
             if (this.selected) {
-                return this.$parent.selectedCandidateIds[this.image.id];
+                return this.image.label;
             }
 
             return null;
         },
         selected() {
-            return this.$parent.selectedCandidateIds.hasOwnProperty(this.image.id);
+            return this.$parent.isSelected(this.image);
         },
         converted() {
-            return this.$parent.convertedCandidateIds.hasOwnProperty(this.image.id);
-        },
-        selectable() {
-            return !this.converted;
+            return this.$parent.isConverted(this.image);
         },
         classObject() {
             return {
@@ -80,7 +72,9 @@ biigle.$component('maia.components.candidatesImageGridImage', {
             return this.image.uuid;
         },
         urlTemplate() {
+            // Usually this would be set in the created function but in this special
+            // case this is not possible.
             return biigle.$require('maia.acUrlTemplate');
         },
     },
-});
+};
