+++ src/resources/assets/js/components/proposalsImageGrid.js
@@ -1,12 +1,21 @@
+import Image from './proposalsImageGridImage';
+import {ImageGrid} from '../import';
+
 /**
  * A letiant of the image grid used for the selection of MAIA training proposals.
  *
  * @type {Object}
  */
-biigle.$component('maia.components.proposalsImageGrid', {
-    mixins: [biigle.$require('volumes.components.imageGrid')],
+export default {
+    mixins: [ImageGrid],
+    template: `<div class="image-grid" @wheel.prevent="scroll">
+        <div class="image-grid__images" ref="images">
+            <image-grid-image v-for="image in displayedImages" :key="image.id" :image="image" :empty-url="emptyUrl" :selectable="selectable" :selected-fade="selectable" :small-icon="!selectable" :selected-icon="selectedIcon" @select="emitSelect"></image-grid-image>
+        </div>
+        <image-grid-progress v-if="canScroll" :progress="progress" @top="jumpToStart" @prev-page="reversePage" @prev-row="reverseRow" @jump="jumpToPercent" @next-row="advanceRow" @next-page="advancePage" @bottom="jumpToEnd"></image-grid-progress>
+    </div>`,
     components: {
-        imageGridImage: biigle.$require('maia.components.proposalsImageGridImage'),
+        imageGridImage: Image,
     },
     props: {
         selectedProposalIds: {
@@ -14,4 +23,4 @@ biigle.$component('maia.components.proposalsImageGrid', {
             required: true,
         },
     },
-});
+};
