+++ src/resources/assets/js/components/proposalsImageGridImage.js
@@ -1,28 +1,23 @@
+import {AnnotationPatch} from '../import';
+import {ImageGridImage} from '../import';
+
 /**
  * A letiant of the image grid image used for the selection of MAIA training proposals.
  *
  * @type {Object}
  */
-biigle.$component('maia.components.proposalsImageGridImage', {
+export default {
     mixins: [
-        biigle.$require('volumes.components.imageGridImage'),
-        biigle.$require('largo.mixins.annotationPatch'),
+        ImageGridImage,
+        AnnotationPatch,
     ],
-    template: '<figure class="image-grid__image" :class="classObject" :title="title">' +
-        '<div v-if="showIcon" class="image-icon">' +
-            '<i class="fas" :class="iconClass"></i>' +
-        '</div>' +
-        '<img @click="toggleSelect" :src="url || emptyUrl" @error="showEmptyImage">' +
-        '<div v-if="showAnnotationLink" class="image-buttons">' +
-            '<a :href="showAnnotationLink" target="_blank" class="image-button" title="Show the annotation in the annotation tool">' +
-                '<span class="fa fa-external-link-square-alt" aria-hidden="true"></span>' +
-            '</a>' +
-        '</div>' +
-    '</figure>',
+    template: `<figure class="image-grid__image" :class="classObject" :title="title">
+        <div v-if="showIcon" class="image-icon">
+            <i class="fas" :class="iconClass"></i>
+        </div>
+        <img @click="toggleSelect" :src="url || emptyUrl" @error="showEmptyImage">
+    </figure>`,
     computed: {
-        showAnnotationLink() {
-            return false;
-        },
         selected() {
             return this.$parent.selectedProposalIds.hasOwnProperty(this.image.id);
         },
@@ -31,15 +26,6 @@ biigle.$component('maia.components.proposalsImageGridImage', {
                 return this.selected ? 'Unselect as interesting' : 'Select as interesting';
             }
         },
-        // Show the small icon when the job is no longer in training proposal state.
-        smallIcon() {
-            return !this.selectable;
-        },
-        // Do not fade selected images when the job is no longer in training proposal
-        // state.
-        selectedFade() {
-            return this.selectable;
-        },
         id() {
             return this.image.id;
         },
@@ -47,7 +33,9 @@ biigle.$component('maia.components.proposalsImageGridImage', {
             return this.image.uuid;
         },
         urlTemplate() {
+            // Usually this would be set in the created function but in this special
+            // case this is not possible.
             return biigle.$require('maia.tpUrlTemplate');
         },
     },
-});
+};
