+++ src/resources/assets/js/components/refineCandidatesCanvas.js
@@ -1,10 +1,17 @@
+import Collection from 'ol/Collection';
+import OlObject from 'ol/Object';
+import RefineCanvas from './refineCanvas';
+import VectorLayer from 'ol/layer/Vector';
+import VectorSource from 'ol/source/Vector';
+import {StylesStore} from '../import';
+
 /**
  * A letiant of the annotation canvas used for the refinement of annotation candidates.
  *
  * @type {Object}
  */
-biigle.$component('maia.components.refineCandidatesCanvas', {
-    mixins: [biigle.$require('maia.components.refineCanvas')],
+export default {
+    mixins: [RefineCanvas],
     props: {
         convertedAnnotations: {
             type: Array,
@@ -15,22 +22,22 @@ biigle.$component('maia.components.refineCandidatesCanvas', {
     },
     methods: {
         createConvertedAnnotationsLayer() {
-            this.convertedAnnotationFeatures = new ol.Collection();
-            this.convertedAnnotationSource = new ol.source.Vector({
+            this.convertedAnnotationFeatures = new Collection();
+            this.convertedAnnotationSource = new VectorSource({
                 features: this.convertedAnnotationFeatures
             });
 
-            let fakeFeature = new ol.Object();
+            let fakeFeature = new OlObject();
             fakeFeature.set('color', '999999');
 
-            this.convertedAnnotationLayer = new ol.layer.Vector({
+            this.convertedAnnotationLayer = new VectorLayer({
                 source: this.convertedAnnotationSource,
                 // Should be below unselected candidates which are at index 99. Else
                 // the attach label interaction doesn't work.
                 zIndex: 98,
                 updateWhileAnimating: true,
                 updateWhileInteracting: true,
-                style: biigle.$require('annotations.stores.styles').features(fakeFeature),
+                style: StylesStore.features(fakeFeature),
             });
         },
     },
@@ -43,4 +50,4 @@ biigle.$component('maia.components.refineCandidatesCanvas', {
         this.createConvertedAnnotationsLayer();
         this.map.addLayer(this.convertedAnnotationLayer);
     },
-});
+};
