+++ src/resources/assets/js/import.js
@@ -1,3 +1,15 @@
-export let LoaderMixin = biigle.$require('core.mixins.loader');
-export let LabelTypeahead = biigle.$require('labelTrees.components.labelTypeahead');
+export let AnnotationCanvas = biigle.$require('annotations.components.annotationCanvas');
+export let AnnotationPatch = biigle.$require('largo.mixins.annotationPatch');
+export let AttachLabelInteraction = biigle.$require('annotations.ol.AttachLabelInteraction');
 export let handleErrorResponse = biigle.$require('messages').handleErrorResponse;
+export let ImageGrid = biigle.$require('volumes.components.imageGrid');
+export let ImageGridImage = biigle.$require('volumes.components.imageGridImage');
+export let ImagesStore = biigle.$require('annotations.stores.images');
+export let Keyboard = biigle.$require('keyboard');
+export let LabelTrees = biigle.$require('labelTrees.components.labelTrees');
+export let LabelTypeahead = biigle.$require('labelTrees.components.labelTypeahead');
+export let LoaderMixin = biigle.$require('core.mixins.loader');
+export let Messages = biigle.$require('messages');
+export let Sidebar = biigle.$require('core.components.sidebar');
+export let SidebarTab = biigle.$require('core.components.sidebarTab');
+export let StylesStore = biigle.$require('annotations.stores.styles');
