+++ src/resources/assets/js/main.js
@@ -1,3 +1,5 @@
 import Form from './form';
+import Show from './show';
 
 biigle.$mount('maia-job-form', Form);
+biigle.$mount('maia-show-container', Show);
